+++ components/moca/index.tsx
@@ -7,7 +7,7 @@ import Bullet from './Bullet';
 import avatar1 from '../../public/images/avatar1.png';
 
 import styles from './index.module.less';
-import { fetchNftCounts } from '../../dataverse/apis/report';
+import { useCurateCountHook } from '../../dataverse/hooks/useDataverseHooks';
 
 const dataList = [
   {
@@ -435,28 +435,6 @@ export default function Moca({ backCall }) {
   const con = classnames('flex justify-center items-center flex-wrap', styles.con);
   const [line, setLine] = useState(0);
   const [wid, setWid] = useState(0);
-  const [did, setDid] = useState('');
-  const [authenticateLoading, setAuthenticateLoading] = useState(false);
-  const [likeLoading, setLikeLoading] = useState(false);
-  const [likeCountsRepData, setLikeCountsRepData] = useState([]);
-  const [isLikedLists, setIsLikedLists] = useState<
-    {
-      url: string;
-      author: string[];
-      twitter: string[];
-      platform: string;
-      platformLink: string;
-      title: string;
-      type: string;
-      desc: string;
-      mcp?: string;
-      chain?: string;
-      contract?: string;
-      tokenId?: string;
-      nftLink?:string;
-      liked?: boolean;
-    }[]
-  >(lists);
 
   // 发送弹幕
   useEffect(() => {
@@ -476,21 +454,7 @@ export default function Moca({ backCall }) {
     }
   }, []);
 
-  useEffect(() => {
-    (async () => {
-      const countsRep = await fetchNftCounts(lists);
-      const likeList = localStorage.getItem('likedList');
-      if (likeList) {
-        const likeListParse = JSON.parse(likeList) as string[];
-        isLikedLists.map((el) => {
-          el.liked = likeListParse.includes(el.nftLink || el.platformLink);
-          return el;
-        });
-        setIsLikedLists(isLikedLists);
-      }
-      setLikeCountsRepData(countsRep.data.data);
-    })();
-  }, []);
+  const { likeCountsRepData, isLikedLists, states } = useCurateCountHook(lists);
 
   return (
     <div className={wrap} id="screen">
@@ -524,14 +488,12 @@ export default function Moca({ backCall }) {
                   <MocaCard
                     {...item}
                     key={index}
-                    likeCount={likeCountsRepData.length > 0 ? likeCountsRepData[index].count : 0}
-                    liked={isLikedLists[index].liked}
-                    did={did}
-                    authenticateLoading={authenticateLoading}
-                    likeLoading={likeLoading}
-                    setDid={setDid}
-                    setAuthenticateLoading={setAuthenticateLoading}
-                    setLikeLoading={setLikeLoading}
+                    data={{
+                      ...item,
+                      likeCount: likeCountsRepData.length > 0 ? likeCountsRepData[index].count : 0,
+                      liked: isLikedLists[index].liked,
+                    }}
+                    states={states}
                   />
                 );
               })}
