--- dataverse/apis/artifacts/SyntheticLoot.ts
@@ -1,325 +1,0 @@
-export const Synthetic_LOOT = [
-    {
-        "inputs": [
-            {
-                "internalType": "address",
-                "name": "walletAddress",
-                "type": "address"
-            }
-        ],
-        "name": "chestComponents",
-        "outputs": [
-            {
-                "internalType": "uint256[5]",
-                "name": "",
-                "type": "uint256[5]"
-            }
-        ],
-        "stateMutability": "view",
-        "type": "function"
-    },
-    {
-        "inputs": [
-            {
-                "internalType": "address",
-                "name": "walletAddress",
-                "type": "address"
-            }
-        ],
-        "name": "footComponents",
-        "outputs": [
-            {
-                "internalType": "uint256[5]",
-                "name": "",
-                "type": "uint256[5]"
-            }
-        ],
-        "stateMutability": "view",
-        "type": "function"
-    },
-    {
-        "inputs": [
-            {
-                "internalType": "address",
-                "name": "walletAddress",
-                "type": "address"
-            }
-        ],
-        "name": "getChest",
-        "outputs": [
-            {
-                "internalType": "string",
-                "name": "",
-                "type": "string"
-            }
-        ],
-        "stateMutability": "view",
-        "type": "function"
-    },
-    {
-        "inputs": [
-            {
-                "internalType": "address",
-                "name": "walletAddress",
-                "type": "address"
-            }
-        ],
-        "name": "getFoot",
-        "outputs": [
-            {
-                "internalType": "string",
-                "name": "",
-                "type": "string"
-            }
-        ],
-        "stateMutability": "view",
-        "type": "function"
-    },
-    {
-        "inputs": [
-            {
-                "internalType": "address",
-                "name": "walletAddress",
-                "type": "address"
-            }
-        ],
-        "name": "getHand",
-        "outputs": [
-            {
-                "internalType": "string",
-                "name": "",
-                "type": "string"
-            }
-        ],
-        "stateMutability": "view",
-        "type": "function"
-    },
-    {
-        "inputs": [
-            {
-                "internalType": "address",
-                "name": "walletAddress",
-                "type": "address"
-            }
-        ],
-        "name": "getHead",
-        "outputs": [
-            {
-                "internalType": "string",
-                "name": "",
-                "type": "string"
-            }
-        ],
-        "stateMutability": "view",
-        "type": "function"
-    },
-    {
-        "inputs": [
-            {
-                "internalType": "address",
-                "name": "walletAddress",
-                "type": "address"
-            }
-        ],
-        "name": "getNeck",
-        "outputs": [
-            {
-                "internalType": "string",
-                "name": "",
-                "type": "string"
-            }
-        ],
-        "stateMutability": "view",
-        "type": "function"
-    },
-    {
-        "inputs": [
-            {
-                "internalType": "address",
-                "name": "walletAddress",
-                "type": "address"
-            }
-        ],
-        "name": "getRing",
-        "outputs": [
-            {
-                "internalType": "string",
-                "name": "",
-                "type": "string"
-            }
-        ],
-        "stateMutability": "view",
-        "type": "function"
-    },
-    {
-        "inputs": [
-            {
-                "internalType": "address",
-                "name": "walletAddress",
-                "type": "address"
-            }
-        ],
-        "name": "getWaist",
-        "outputs": [
-            {
-                "internalType": "string",
-                "name": "",
-                "type": "string"
-            }
-        ],
-        "stateMutability": "view",
-        "type": "function"
-    },
-    {
-        "inputs": [
-            {
-                "internalType": "address",
-                "name": "walletAddress",
-                "type": "address"
-            }
-        ],
-        "name": "getWeapon",
-        "outputs": [
-            {
-                "internalType": "string",
-                "name": "",
-                "type": "string"
-            }
-        ],
-        "stateMutability": "view",
-        "type": "function"
-    },
-    {
-        "inputs": [
-            {
-                "internalType": "address",
-                "name": "walletAddress",
-                "type": "address"
-            }
-        ],
-        "name": "handComponents",
-        "outputs": [
-            {
-                "internalType": "uint256[5]",
-                "name": "",
-                "type": "uint256[5]"
-            }
-        ],
-        "stateMutability": "view",
-        "type": "function"
-    },
-    {
-        "inputs": [
-            {
-                "internalType": "address",
-                "name": "walletAddress",
-                "type": "address"
-            }
-        ],
-        "name": "headComponents",
-        "outputs": [
-            {
-                "internalType": "uint256[5]",
-                "name": "",
-                "type": "uint256[5]"
-            }
-        ],
-        "stateMutability": "view",
-        "type": "function"
-    },
-    {
-        "inputs": [
-            {
-                "internalType": "address",
-                "name": "walletAddress",
-                "type": "address"
-            }
-        ],
-        "name": "neckComponents",
-        "outputs": [
-            {
-                "internalType": "uint256[5]",
-                "name": "",
-                "type": "uint256[5]"
-            }
-        ],
-        "stateMutability": "view",
-        "type": "function"
-    },
-    {
-        "inputs": [
-            {
-                "internalType": "address",
-                "name": "walletAddress",
-                "type": "address"
-            }
-        ],
-        "name": "ringComponents",
-        "outputs": [
-            {
-                "internalType": "uint256[5]",
-                "name": "",
-                "type": "uint256[5]"
-            }
-        ],
-        "stateMutability": "view",
-        "type": "function"
-    },
-    {
-        "inputs": [
-            {
-                "internalType": "address",
-                "name": "walletAddress",
-                "type": "address"
-            }
-        ],
-        "name": "tokenURI",
-        "outputs": [
-            {
-                "internalType": "string",
-                "name": "",
-                "type": "string"
-            }
-        ],
-        "stateMutability": "view",
-        "type": "function"
-    },
-    {
-        "inputs": [
-            {
-                "internalType": "address",
-                "name": "walletAddress",
-                "type": "address"
-            }
-        ],
-        "name": "waistComponents",
-        "outputs": [
-            {
-                "internalType": "uint256[5]",
-                "name": "",
-                "type": "uint256[5]"
-            }
-        ],
-        "stateMutability": "view",
-        "type": "function"
-    },
-    {
-        "inputs": [
-            {
-                "internalType": "address",
-                "name": "walletAddress",
-                "type": "address"
-            }
-        ],
-        "name": "weaponComponents",
-        "outputs": [
-            {
-                "internalType": "uint256[5]",
-                "name": "",
-                "type": "uint256[5]"
-            }
-        ],
-        "stateMutability": "view",
-        "type": "function"
-    }
-]
