--- dataverse/apis/artifacts/layerMapping.ts
@@ -1,1842 +1,0 @@
-export const LAYER_MAPPING = {
-    "VERSION": "0.0.1",
-    "weapon": {
-        "prefix": "weapon/prefix/prefix.png",
-        "suffix": {
-            "Grave Wand": {
-                "of Skill": "weapon/suffix/grave_wand_of_skill.png",
-                "of Protection": "weapon/suffix/grave_wand_of_protection.png",
-                "of Brilliance": "weapon/suffix/grave_wand_of_brilliance.png",
-                "of Giants": "weapon/suffix/grave_wand_of_giants.png",
-                "of Vitriol": "weapon/suffix/grave_wand_of_vitriol.png",
-                "of Detection": "weapon/suffix/grave_wand_of_detection.png",
-                "of the Twins": "weapon/suffix/grave_wand_of_the_twins.png",
-                "of Rage": "weapon/suffix/grave_wand_of_rage.png"
-            },
-            "Bone Wand": {
-                "of Perfection": "weapon/suffix/bone_wand_of_perfection.png",
-                "of Reflection": "weapon/suffix/bone_wand_of_reflection.png",
-                "of Anger": "weapon/suffix/bone_wand_of_anger.png",
-                "of Power": "weapon/suffix/bone_wand_of_power.png",
-                "of Titans": "weapon/suffix/bone_wand_of_titans.png",
-                "of the Fox": "weapon/suffix/bone_wand_of_the_fox.png",
-                "of Enlightenment": "weapon/suffix/bone_wand_of_enlightenment.png",
-                "of Fury": "weapon/suffix/bone_wand_of_fury.png"
-            },
-            "Katana": {
-                "of Rage": "weapon/suffix/katana_of_rage.png",
-                "of Protection": "weapon/suffix/katana_of_protection.png",
-                "of Skill": "weapon/suffix/katana_of_skill.png",
-                "of Brilliance": "weapon/suffix/katana_of_brilliance.png",
-                "of the Twins": "weapon/suffix/katana_of_the_twins.png",
-                "of Detection": "weapon/suffix/katana_of_detection.png",
-                "of Giants": "weapon/suffix/katana_of_giants.png",
-                "of Vitriol": "weapon/suffix/katana_of_vitriol.png"
-            },
-            "Scimitar": {
-                "of Detection": "weapon/suffix/scimitar_of_detection.png",
-                "of Brilliance": "weapon/suffix/scimitar_of_brilliance.png",
-                "of Vitriol": "weapon/suffix/scimitar_of_vitriol.png",
-                "of Rage": "weapon/suffix/scimitar_of_rage.png",
-                "of Giants": "weapon/suffix/scimitar_of_giants.png",
-                "of Skill": "weapon/suffix/scimitar_of_skill.png",
-                "of Protection": "weapon/suffix/scimitar_of_protection.png",
-                "of the Twins": "weapon/suffix/scimitar_of_the_twins.png"
-            },
-            "Maul": {
-                "of Reflection": "weapon/suffix/maul_of_reflection.png",
-                "of Titans": "weapon/suffix/maul_of_titans.png",
-                "of Enlightenment": "weapon/suffix/maul_of_enlightenment.png",
-                "of Anger": "weapon/suffix/maul_of_anger.png",
-                "of Fury": "weapon/suffix/maul_of_fury.png",
-                "of Power": "weapon/suffix/maul_of_power.png",
-                "of the Fox": "weapon/suffix/maul_of_the_fox.png",
-                "of Perfection": "weapon/suffix/maul_of_perfection.png"
-            },
-            "Long Sword": {
-                "of Titans": "weapon/suffix/long_sword_of_titans.png",
-                "of Perfection": "weapon/suffix/long_sword_of_perfection.png",
-                "of Power": "weapon/suffix/long_sword_of_power.png",
-                "of Fury": "weapon/suffix/long_sword_of_fury.png",
-                "of the Fox": "weapon/suffix/long_sword_of_the_fox.png",
-                "of Reflection": "weapon/suffix/long_sword_of_reflection.png",
-                "of Enlightenment": "weapon/suffix/long_sword_of_enlightenment.png",
-                "of Anger": "weapon/suffix/long_sword_of_anger.png"
-            },
-            "Ghost Wand": {
-                "of the Fox": "weapon/suffix/ghost_wand_of_the_fox.png",
-                "of Fury": "weapon/suffix/ghost_wand_of_fury.png",
-                "of Titans": "weapon/suffix/ghost_wand_of_titans.png",
-                "of Reflection": "weapon/suffix/ghost_wand_of_reflection.png",
-                "of Anger": "weapon/suffix/ghost_wand_of_anger.png",
-                "of Enlightenment": "weapon/suffix/ghost_wand_of_enlightenment.png",
-                "of Power": "weapon/suffix/ghost_wand_of_power.png",
-                "of Perfection": "weapon/suffix/ghost_wand_of_perfection.png"
-            },
-            "Short Sword": {
-                "of Rage": "weapon/suffix/short_sword_of_rage.png",
-                "of Brilliance": "weapon/suffix/short_sword_of_brilliance.png",
-                "of Giants": "weapon/suffix/short_sword_of_giants.png",
-                "of Skill": "weapon/suffix/short_sword_of_skill.png",
-                "of Detection": "weapon/suffix/short_sword_of_detection.png",
-                "of Vitriol": "weapon/suffix/short_sword_of_vitriol.png",
-                "of Protection": "weapon/suffix/short_sword_of_protection.png",
-                "of the Twins": "weapon/suffix/short_sword_of_the_twins.png"
-            },
-            "Warhammer": {
-                "of the Fox": "weapon/suffix/warhammer_of_the_fox.png",
-                "of Perfection": "weapon/suffix/warhammer_of_perfection.png",
-                "of Anger": "weapon/suffix/warhammer_of_anger.png",
-                "of Enlightenment": "weapon/suffix/warhammer_of_enlightenment.png",
-                "of Reflection": "weapon/suffix/warhammer_of_reflection.png",
-                "of Fury": "weapon/suffix/warhammer_of_fury.png",
-                "of Power": "weapon/suffix/warhammer_of_power.png",
-                "of Titans": "weapon/suffix/warhammer_of_titans.png"
-            },
-            "Chronicle": {
-                "of the Twins": "weapon/suffix/chronicle_of_the_twins.png",
-                "of Vitriol": "weapon/suffix/chronicle_of_vitriol.png",
-                "of Brilliance": "weapon/suffix/chronicle_of_brilliance.png",
-                "of Skill": "weapon/suffix/chronicle_of_skill.png",
-                "of Detection": "weapon/suffix/chronicle_of_detection.png",
-                "of Giants": "weapon/suffix/chronicle_of_giants.png",
-                "of Rage": "weapon/suffix/chronicle_of_rage.png",
-                "of Protection": "weapon/suffix/chronicle_of_protection.png"
-            },
-            "Book": {
-                "of Giants": "weapon/suffix/book_of_giants.png",
-                "of Skill": "weapon/suffix/book_of_skill.png",
-                "of Rage": "weapon/suffix/book_of_rage.png",
-                "of the Twins": "weapon/suffix/book_of_the_twins.png",
-                "of Protection": "weapon/suffix/book_of_protection.png",
-                "of Brilliance": "weapon/suffix/book_of_brilliance.png",
-                "of Vitriol": "weapon/suffix/book_of_vitriol.png",
-                "of Detection": "weapon/suffix/book_of_detection.png"
-            },
-            "Tome": {
-                "of Anger": "weapon/suffix/tome_of_anger.png",
-                "of Fury": "weapon/suffix/tome_of_fury.png",
-                "of the Fox": "weapon/suffix/tome_of_the_fox.png",
-                "of Power": "weapon/suffix/tome_of_power.png",
-                "of Titans": "weapon/suffix/tome_of_titans.png",
-                "of Enlightenment": "weapon/suffix/tome_of_enlightenment.png",
-                "of Perfection": "weapon/suffix/tome_of_perfection.png",
-                "of Reflection": "weapon/suffix/tome_of_reflection.png"
-            },
-            "Quarterstaff": {
-                "of Skill": "weapon/suffix/quarterstaff_of_skill.png",
-                "of Vitriol": "weapon/suffix/quarterstaff_of_vitriol.png",
-                "of Giants": "weapon/suffix/quarterstaff_of_giants.png",
-                "of Protection": "weapon/suffix/quarterstaff_of_protection.png",
-                "of Detection": "weapon/suffix/quarterstaff_of_detection.png",
-                "of the Twins": "weapon/suffix/quarterstaff_of_the_twins.png",
-                "of Rage": "weapon/suffix/quarterstaff_of_rage.png",
-                "of Brilliance": "weapon/suffix/quarterstaff_of_brilliance.png"
-            },
-            "Grimoire": {
-                "of Enlightenment": "weapon/suffix/grimoire_of_enlightenment.png",
-                "of Power": "weapon/suffix/grimoire_of_power.png",
-                "of Fury": "weapon/suffix/grimoire_of_fury.png",
-                "of the Fox": "weapon/suffix/grimoire_of_the_fox.png",
-                "of Reflection": "weapon/suffix/grimoire_of_reflection.png",
-                "of Perfection": "weapon/suffix/grimoire_of_perfection.png",
-                "of Titans": "weapon/suffix/grimoire_of_titans.png",
-                "of Anger": "weapon/suffix/grimoire_of_anger.png"
-            },
-            "Falchion": {
-                "of Reflection": "weapon/suffix/falchion_of_reflection.png",
-                "of Power": "weapon/suffix/falchion_of_power.png",
-                "of Titans": "weapon/suffix/falchion_of_titans.png",
-                "of Fury": "weapon/suffix/falchion_of_fury.png",
-                "of Anger": "weapon/suffix/falchion_of_anger.png",
-                "of Enlightenment": "weapon/suffix/falchion_of_enlightenment.png",
-                "of the Fox": "weapon/suffix/falchion_of_the_fox.png",
-                "of Perfection": "weapon/suffix/falchion_of_perfection.png"
-            },
-            "Mace": {
-                "of Rage": "weapon/suffix/mace_of_rage.png",
-                "of Giants": "weapon/suffix/mace_of_giants.png",
-                "of Skill": "weapon/suffix/mace_of_skill.png",
-                "of Detection": "weapon/suffix/mace_of_detection.png",
-                "of Vitriol": "weapon/suffix/mace_of_vitriol.png",
-                "of the Twins": "weapon/suffix/mace_of_the_twins.png",
-                "of Protection": "weapon/suffix/mace_of_protection.png",
-                "of Brilliance": "weapon/suffix/mace_of_brilliance.png"
-            },
-            "Club": {
-                "of Fury": "weapon/suffix/club_of_fury.png",
-                "of Perfection": "weapon/suffix/club_of_perfection.png",
-                "of Enlightenment": "weapon/suffix/club_of_enlightenment.png",
-                "of Reflection": "weapon/suffix/club_of_reflection.png",
-                "of Titans": "weapon/suffix/club_of_titans.png",
-                "of Anger": "weapon/suffix/club_of_anger.png",
-                "of the Fox": "weapon/suffix/club_of_the_fox.png",
-                "of Power": "weapon/suffix/club_of_power.png"
-            },
-            "Wand": {
-                "of Skill": "weapon/suffix/wand_of_skill.png",
-                "of the Twins": "weapon/suffix/wand_of_the_twins.png",
-                "of Protection": "weapon/suffix/wand_of_protection.png",
-                "of Vitriol": "weapon/suffix/wand_of_vitriol.png",
-                "of Giants": "weapon/suffix/wand_of_giants.png",
-                "of Brilliance": "weapon/suffix/wand_of_brilliance.png",
-                "of Detection": "weapon/suffix/wand_of_detection.png",
-                "of Rage": "weapon/suffix/wand_of_rage.png"
-            }
-        },
-        "plus_one": "weapon/plus_one/plus_one.png",
-        "name": {
-            "Grave Wand": "weapon/name/grave_wand.png",
-            "Bone Wand": "weapon/name/bone_wand.png",
-            "Katana": "weapon/name/katana.png",
-            "Scimitar": "weapon/name/scimitar.png",
-            "Maul": "weapon/name/maul.png",
-            "Long Sword": "weapon/name/long_sword.png",
-            "Ghost Wand": "weapon/name/ghost_wand.png",
-            "Short Sword": "weapon/name/short_sword.png",
-            "Warhammer": "weapon/name/warhammer.png",
-            "Chronicle": "weapon/name/chronicle.png",
-            "Book": "weapon/name/book.png",
-            "Tome": "weapon/name/tome.png",
-            "Quarterstaff": "weapon/name/quarterstaff.png",
-            "Grimoire": "weapon/name/grimoire.png",
-            "Falchion": "weapon/name/falchion.png",
-            "Mace": "weapon/name/mace.png",
-            "Club": "weapon/name/club.png",
-            "Wand": "weapon/name/wand.png"
-        }
-    },
-    "chest": {
-        "prefix": "chest/prefix/prefix.png",
-        "suffix": {
-            "Hard Leather Armor": {
-                "of Skill": "chest/suffix/hard_leather_armor_of_skill.png",
-                "of Giants": "chest/suffix/hard_leather_armor_of_giants.png",
-                "of Protection": "chest/suffix/hard_leather_armor_of_protection.png",
-                "of Rage": "chest/suffix/hard_leather_armor_of_rage.png",
-                "of Perfection": "chest/suffix/hard_leather_armor_of_perfection.png",
-                "of the Fox": "chest/suffix/hard_leather_armor_of_the_fox.png",
-                "of Vitriol": "chest/suffix/hard_leather_armor_of_vitriol.png",
-                "of Brilliance": "chest/suffix/hard_leather_armor_of_brilliance.png",
-                "of Detection": "chest/suffix/hard_leather_armor_of_detection.png",
-                "of Anger": "chest/suffix/hard_leather_armor_of_anger.png",
-                "of Reflection": "chest/suffix/hard_leather_armor_of_reflection.png",
-                "of Power": "chest/suffix/hard_leather_armor_of_power.png",
-                "of Fury": "chest/suffix/hard_leather_armor_of_fury.png",
-                "of Titans": "chest/suffix/hard_leather_armor_of_titans.png",
-                "of the Twins": "chest/suffix/hard_leather_armor_of_the_twins.png",
-                "of Enlightenment": "chest/suffix/hard_leather_armor_of_enlightenment.png"
-            },
-            "Silk Robe": {
-                "of Fury": "chest/suffix/silk_robe_of_fury.png",
-                "of Anger": "chest/suffix/silk_robe_of_anger.png",
-                "of Titans": "chest/suffix/silk_robe_of_titans.png",
-                "of Brilliance": "chest/suffix/silk_robe_of_brilliance.png",
-                "of Perfection": "chest/suffix/silk_robe_of_perfection.png",
-                "of Protection": "chest/suffix/silk_robe_of_protection.png",
-                "of Vitriol": "chest/suffix/silk_robe_of_vitriol.png",
-                "of Giants": "chest/suffix/silk_robe_of_giants.png",
-                "of the Twins": "chest/suffix/silk_robe_of_the_twins.png",
-                "of Rage": "chest/suffix/silk_robe_of_rage.png",
-                "of Power": "chest/suffix/silk_robe_of_power.png",
-                "of Skill": "chest/suffix/silk_robe_of_skill.png",
-                "of Detection": "chest/suffix/silk_robe_of_detection.png",
-                "of Enlightenment": "chest/suffix/silk_robe_of_enlightenment.png",
-                "of the Fox": "chest/suffix/silk_robe_of_the_fox.png",
-                "of Reflection": "chest/suffix/silk_robe_of_reflection.png"
-            },
-            "Ornate Chestplate": {
-                "of Skill": "chest/suffix/ornate_chestplate_of_skill.png",
-                "of Anger": "chest/suffix/ornate_chestplate_of_anger.png",
-                "of Detection": "chest/suffix/ornate_chestplate_of_detection.png",
-                "of Giants": "chest/suffix/ornate_chestplate_of_giants.png",
-                "of Power": "chest/suffix/ornate_chestplate_of_power.png",
-                "of Fury": "chest/suffix/ornate_chestplate_of_fury.png",
-                "of Brilliance": "chest/suffix/ornate_chestplate_of_brilliance.png",
-                "of the Twins": "chest/suffix/ornate_chestplate_of_the_twins.png",
-                "of Enlightenment": "chest/suffix/ornate_chestplate_of_enlightenment.png",
-                "of Protection": "chest/suffix/ornate_chestplate_of_protection.png",
-                "of Rage": "chest/suffix/ornate_chestplate_of_rage.png",
-                "of Reflection": "chest/suffix/ornate_chestplate_of_reflection.png",
-                "of the Fox": "chest/suffix/ornate_chestplate_of_the_fox.png",
-                "of Titans": "chest/suffix/ornate_chestplate_of_titans.png",
-                "of Vitriol": "chest/suffix/ornate_chestplate_of_vitriol.png",
-                "of Perfection": "chest/suffix/ornate_chestplate_of_perfection.png"
-            },
-            "Plate Mail": {
-                "of Titans": "chest/suffix/plate_mail_of_titans.png",
-                "of Detection": "chest/suffix/plate_mail_of_detection.png",
-                "of Fury": "chest/suffix/plate_mail_of_fury.png",
-                "of Perfection": "chest/suffix/plate_mail_of_perfection.png",
-                "of Power": "chest/suffix/plate_mail_of_power.png",
-                "of Reflection": "chest/suffix/plate_mail_of_reflection.png",
-                "of Enlightenment": "chest/suffix/plate_mail_of_enlightenment.png",
-                "of Vitriol": "chest/suffix/plate_mail_of_vitriol.png",
-                "of the Twins": "chest/suffix/plate_mail_of_the_twins.png",
-                "of the Fox": "chest/suffix/plate_mail_of_the_fox.png",
-                "of Giants": "chest/suffix/plate_mail_of_giants.png",
-                "of Brilliance": "chest/suffix/plate_mail_of_brilliance.png",
-                "of Protection": "chest/suffix/plate_mail_of_protection.png",
-                "of Anger": "chest/suffix/plate_mail_of_anger.png",
-                "of Skill": "chest/suffix/plate_mail_of_skill.png",
-                "of Rage": "chest/suffix/plate_mail_of_rage.png"
-            },
-            "Dragonskin Armor": {
-                "of Giants": "chest/suffix/dragonskin_armor_of_giants.png",
-                "of Power": "chest/suffix/dragonskin_armor_of_power.png",
-                "of Reflection": "chest/suffix/dragonskin_armor_of_reflection.png",
-                "of Rage": "chest/suffix/dragonskin_armor_of_rage.png",
-                "of Protection": "chest/suffix/dragonskin_armor_of_protection.png",
-                "of Fury": "chest/suffix/dragonskin_armor_of_fury.png",
-                "of Titans": "chest/suffix/dragonskin_armor_of_titans.png",
-                "of Vitriol": "chest/suffix/dragonskin_armor_of_vitriol.png",
-                "of Enlightenment": "chest/suffix/dragonskin_armor_of_enlightenment.png",
-                "of Perfection": "chest/suffix/dragonskin_armor_of_perfection.png",
-                "of Anger": "chest/suffix/dragonskin_armor_of_anger.png",
-                "of the Fox": "chest/suffix/dragonskin_armor_of_the_fox.png",
-                "of the Twins": "chest/suffix/dragonskin_armor_of_the_twins.png",
-                "of Detection": "chest/suffix/dragonskin_armor_of_detection.png",
-                "of Brilliance": "chest/suffix/dragonskin_armor_of_brilliance.png",
-                "of Skill": "chest/suffix/dragonskin_armor_of_skill.png"
-            },
-            "Shirt": {
-                "of Giants": "chest/suffix/shirt_of_giants.png",
-                "of Power": "chest/suffix/shirt_of_power.png",
-                "of Anger": "chest/suffix/shirt_of_anger.png",
-                "of Enlightenment": "chest/suffix/shirt_of_enlightenment.png",
-                "of Rage": "chest/suffix/shirt_of_rage.png",
-                "of Brilliance": "chest/suffix/shirt_of_brilliance.png",
-                "of Skill": "chest/suffix/shirt_of_skill.png",
-                "of the Twins": "chest/suffix/shirt_of_the_twins.png",
-                "of Fury": "chest/suffix/shirt_of_fury.png",
-                "of Protection": "chest/suffix/shirt_of_protection.png",
-                "of Reflection": "chest/suffix/shirt_of_reflection.png",
-                "of Titans": "chest/suffix/shirt_of_titans.png",
-                "of Detection": "chest/suffix/shirt_of_detection.png",
-                "of the Fox": "chest/suffix/shirt_of_the_fox.png",
-                "of Perfection": "chest/suffix/shirt_of_perfection.png",
-                "of Vitriol": "chest/suffix/shirt_of_vitriol.png"
-            },
-            "Studded Leather Armor": {
-                "of Rage": "chest/suffix/studded_leather_armor_of_rage.png",
-                "of Protection": "chest/suffix/studded_leather_armor_of_protection.png",
-                "of Vitriol": "chest/suffix/studded_leather_armor_of_vitriol.png",
-                "of Anger": "chest/suffix/studded_leather_armor_of_anger.png",
-                "of Giants": "chest/suffix/studded_leather_armor_of_giants.png",
-                "of Skill": "chest/suffix/studded_leather_armor_of_skill.png",
-                "of Reflection": "chest/suffix/studded_leather_armor_of_reflection.png",
-                "of Perfection": "chest/suffix/studded_leather_armor_of_perfection.png",
-                "of Fury": "chest/suffix/studded_leather_armor_of_fury.png",
-                "of the Twins": "chest/suffix/studded_leather_armor_of_the_twins.png",
-                "of the Fox": "chest/suffix/studded_leather_armor_of_the_fox.png",
-                "of Power": "chest/suffix/studded_leather_armor_of_power.png",
-                "of Enlightenment": "chest/suffix/studded_leather_armor_of_enlightenment.png",
-                "of Detection": "chest/suffix/studded_leather_armor_of_detection.png",
-                "of Titans": "chest/suffix/studded_leather_armor_of_titans.png",
-                "of Brilliance": "chest/suffix/studded_leather_armor_of_brilliance.png"
-            },
-            "Robe": {
-                "of Giants": "chest/suffix/robe_of_giants.png",
-                "of Anger": "chest/suffix/robe_of_anger.png",
-                "of Brilliance": "chest/suffix/robe_of_brilliance.png",
-                "of Power": "chest/suffix/robe_of_power.png",
-                "of Titans": "chest/suffix/robe_of_titans.png",
-                "of the Twins": "chest/suffix/robe_of_the_twins.png",
-                "of Detection": "chest/suffix/robe_of_detection.png",
-                "of Fury": "chest/suffix/robe_of_fury.png",
-                "of Vitriol": "chest/suffix/robe_of_vitriol.png",
-                "of Reflection": "chest/suffix/robe_of_reflection.png",
-                "of Protection": "chest/suffix/robe_of_protection.png",
-                "of the Fox": "chest/suffix/robe_of_the_fox.png",
-                "of Enlightenment": "chest/suffix/robe_of_enlightenment.png",
-                "of Rage": "chest/suffix/robe_of_rage.png",
-                "of Perfection": "chest/suffix/robe_of_perfection.png",
-                "of Skill": "chest/suffix/robe_of_skill.png"
-            },
-            "Chain Mail": {
-                "of Detection": "chest/suffix/chain_mail_of_detection.png",
-                "of Brilliance": "chest/suffix/chain_mail_of_brilliance.png",
-                "of Protection": "chest/suffix/chain_mail_of_protection.png",
-                "of Perfection": "chest/suffix/chain_mail_of_perfection.png",
-                "of Skill": "chest/suffix/chain_mail_of_skill.png",
-                "of Reflection": "chest/suffix/chain_mail_of_reflection.png",
-                "of Titans": "chest/suffix/chain_mail_of_titans.png",
-                "of Vitriol": "chest/suffix/chain_mail_of_vitriol.png",
-                "of the Fox": "chest/suffix/chain_mail_of_the_fox.png",
-                "of Rage": "chest/suffix/chain_mail_of_rage.png",
-                "of Giants": "chest/suffix/chain_mail_of_giants.png",
-                "of Power": "chest/suffix/chain_mail_of_power.png",
-                "of Anger": "chest/suffix/chain_mail_of_anger.png",
-                "of Fury": "chest/suffix/chain_mail_of_fury.png",
-                "of Enlightenment": "chest/suffix/chain_mail_of_enlightenment.png",
-                "of the Twins": "chest/suffix/chain_mail_of_the_twins.png"
-            },
-            "Demon Husk": {
-                "of Detection": "chest/suffix/demon_husk_of_detection.png",
-                "of Skill": "chest/suffix/demon_husk_of_skill.png",
-                "of Titans": "chest/suffix/demon_husk_of_titans.png",
-                "of the Fox": "chest/suffix/demon_husk_of_the_fox.png",
-                "of Protection": "chest/suffix/demon_husk_of_protection.png",
-                "of Power": "chest/suffix/demon_husk_of_power.png",
-                "of Perfection": "chest/suffix/demon_husk_of_perfection.png",
-                "of the Twins": "chest/suffix/demon_husk_of_the_twins.png",
-                "of Fury": "chest/suffix/demon_husk_of_fury.png",
-                "of Brilliance": "chest/suffix/demon_husk_of_brilliance.png",
-                "of Reflection": "chest/suffix/demon_husk_of_reflection.png",
-                "of Enlightenment": "chest/suffix/demon_husk_of_enlightenment.png",
-                "of Anger": "chest/suffix/demon_husk_of_anger.png",
-                "of Giants": "chest/suffix/demon_husk_of_giants.png",
-                "of Rage": "chest/suffix/demon_husk_of_rage.png",
-                "of Vitriol": "chest/suffix/demon_husk_of_vitriol.png"
-            },
-            "Linen Robe": {
-                "of Rage": "chest/suffix/linen_robe_of_rage.png",
-                "of Protection": "chest/suffix/linen_robe_of_protection.png",
-                "of Power": "chest/suffix/linen_robe_of_power.png",
-                "of Vitriol": "chest/suffix/linen_robe_of_vitriol.png",
-                "of Anger": "chest/suffix/linen_robe_of_anger.png",
-                "of Brilliance": "chest/suffix/linen_robe_of_brilliance.png",
-                "of Skill": "chest/suffix/linen_robe_of_skill.png",
-                "of Reflection": "chest/suffix/linen_robe_of_reflection.png",
-                "of Giants": "chest/suffix/linen_robe_of_giants.png",
-                "of the Fox": "chest/suffix/linen_robe_of_the_fox.png",
-                "of Fury": "chest/suffix/linen_robe_of_fury.png",
-                "of the Twins": "chest/suffix/linen_robe_of_the_twins.png",
-                "of Detection": "chest/suffix/linen_robe_of_detection.png",
-                "of Titans": "chest/suffix/linen_robe_of_titans.png",
-                "of Perfection": "chest/suffix/linen_robe_of_perfection.png",
-                "of Enlightenment": "chest/suffix/linen_robe_of_enlightenment.png"
-            },
-            "Ring Mail": {
-                "of Vitriol": "chest/suffix/ring_mail_of_vitriol.png",
-                "of Giants": "chest/suffix/ring_mail_of_giants.png",
-                "of Enlightenment": "chest/suffix/ring_mail_of_enlightenment.png",
-                "of Fury": "chest/suffix/ring_mail_of_fury.png",
-                "of Detection": "chest/suffix/ring_mail_of_detection.png",
-                "of Power": "chest/suffix/ring_mail_of_power.png",
-                "of Anger": "chest/suffix/ring_mail_of_anger.png",
-                "of the Twins": "chest/suffix/ring_mail_of_the_twins.png",
-                "of Titans": "chest/suffix/ring_mail_of_titans.png",
-                "of Reflection": "chest/suffix/ring_mail_of_reflection.png",
-                "of Rage": "chest/suffix/ring_mail_of_rage.png",
-                "of Perfection": "chest/suffix/ring_mail_of_perfection.png",
-                "of the Fox": "chest/suffix/ring_mail_of_the_fox.png",
-                "of Brilliance": "chest/suffix/ring_mail_of_brilliance.png",
-                "of Skill": "chest/suffix/ring_mail_of_skill.png",
-                "of Protection": "chest/suffix/ring_mail_of_protection.png"
-            },
-            "Leather Armor": {
-                "of Perfection": "chest/suffix/leather_armor_of_perfection.png",
-                "of Protection": "chest/suffix/leather_armor_of_protection.png",
-                "of Enlightenment": "chest/suffix/leather_armor_of_enlightenment.png",
-                "of the Twins": "chest/suffix/leather_armor_of_the_twins.png",
-                "of Power": "chest/suffix/leather_armor_of_power.png",
-                "of Skill": "chest/suffix/leather_armor_of_skill.png",
-                "of Titans": "chest/suffix/leather_armor_of_titans.png",
-                "of the Fox": "chest/suffix/leather_armor_of_the_fox.png",
-                "of Rage": "chest/suffix/leather_armor_of_rage.png",
-                "of Detection": "chest/suffix/leather_armor_of_detection.png",
-                "of Giants": "chest/suffix/leather_armor_of_giants.png",
-                "of Vitriol": "chest/suffix/leather_armor_of_vitriol.png",
-                "of Anger": "chest/suffix/leather_armor_of_anger.png",
-                "of Fury": "chest/suffix/leather_armor_of_fury.png",
-                "of Brilliance": "chest/suffix/leather_armor_of_brilliance.png",
-                "of Reflection": "chest/suffix/leather_armor_of_reflection.png"
-            },
-            "Divine Robe": {
-                "of Rage": "chest/suffix/divine_robe_of_rage.png",
-                "of Power": "chest/suffix/divine_robe_of_power.png",
-                "of Detection": "chest/suffix/divine_robe_of_detection.png",
-                "of Giants": "chest/suffix/divine_robe_of_giants.png",
-                "of Perfection": "chest/suffix/divine_robe_of_perfection.png",
-                "of Reflection": "chest/suffix/divine_robe_of_reflection.png",
-                "of Fury": "chest/suffix/divine_robe_of_fury.png",
-                "of Vitriol": "chest/suffix/divine_robe_of_vitriol.png",
-                "of the Fox": "chest/suffix/divine_robe_of_the_fox.png",
-                "of Skill": "chest/suffix/divine_robe_of_skill.png",
-                "of Brilliance": "chest/suffix/divine_robe_of_brilliance.png",
-                "of Titans": "chest/suffix/divine_robe_of_titans.png",
-                "of Protection": "chest/suffix/divine_robe_of_protection.png",
-                "of Enlightenment": "chest/suffix/divine_robe_of_enlightenment.png",
-                "of the Twins": "chest/suffix/divine_robe_of_the_twins.png",
-                "of Anger": "chest/suffix/divine_robe_of_anger.png"
-            },
-            "Holy Chestplate": {
-                "of Titans": "chest/suffix/holy_chestplate_of_titans.png",
-                "of Perfection": "chest/suffix/holy_chestplate_of_perfection.png",
-                "of Skill": "chest/suffix/holy_chestplate_of_skill.png",
-                "of Vitriol": "chest/suffix/holy_chestplate_of_vitriol.png",
-                "of the Twins": "chest/suffix/holy_chestplate_of_the_twins.png",
-                "of Protection": "chest/suffix/holy_chestplate_of_protection.png",
-                "of Rage": "chest/suffix/holy_chestplate_of_rage.png",
-                "of Reflection": "chest/suffix/holy_chestplate_of_reflection.png",
-                "of the Fox": "chest/suffix/holy_chestplate_of_the_fox.png",
-                "of Power": "chest/suffix/holy_chestplate_of_power.png",
-                "of Brilliance": "chest/suffix/holy_chestplate_of_brilliance.png",
-                "of Anger": "chest/suffix/holy_chestplate_of_anger.png",
-                "of Detection": "chest/suffix/holy_chestplate_of_detection.png",
-                "of Enlightenment": "chest/suffix/holy_chestplate_of_enlightenment.png",
-                "of Fury": "chest/suffix/holy_chestplate_of_fury.png",
-                "of Giants": "chest/suffix/holy_chestplate_of_giants.png"
-            }
-        },
-        "plus_one": "chest/plus_one/plus_one.png",
-        "name": {
-            "Hard Leather Armor": "chest/name/hard_leather_armor.png",
-            "Silk Robe": "chest/name/silk_robe.png",
-            "Ornate Chestplate": "chest/name/ornate_chestplate.png",
-            "Plate Mail": "chest/name/plate_mail.png",
-            "Dragonskin Armor": "chest/name/dragonskin_armor.png",
-            "Shirt": "chest/name/shirt.png",
-            "Studded Leather Armor": "chest/name/studded_leather_armor.png",
-            "Robe": "chest/name/robe.png",
-            "Chain Mail": "chest/name/chain_mail.png",
-            "Demon Husk": "chest/name/demon_husk.png",
-            "Linen Robe": "chest/name/linen_robe.png",
-            "Ring Mail": "chest/name/ring_mail.png",
-            "Leather Armor": "chest/name/leather_armor.png",
-            "Divine Robe": "chest/name/divine_robe.png",
-            "Holy Chestplate": "chest/name/holy_chestplate.png"
-        }
-    },
-    "head": {
-        "prefix": "head/prefix/prefix.png",
-        "suffix": {
-            "Divine Hood": {
-                "of Protection": "head/suffix/divine_hood_of_protection.png",
-                "of Reflection": "head/suffix/divine_hood_of_reflection.png",
-                "of Anger": "head/suffix/divine_hood_of_anger.png",
-                "of Vitriol": "head/suffix/divine_hood_of_vitriol.png",
-                "of Titans": "head/suffix/divine_hood_of_titans.png",
-                "of Detection": "head/suffix/divine_hood_of_detection.png",
-                "of Brilliance": "head/suffix/divine_hood_of_brilliance.png",
-                "of Fury": "head/suffix/divine_hood_of_fury.png",
-                "of Power": "head/suffix/divine_hood_of_power.png",
-                "of Skill": "head/suffix/divine_hood_of_skill.png",
-                "of Giants": "head/suffix/divine_hood_of_giants.png",
-                "of the Twins": "head/suffix/divine_hood_of_the_twins.png",
-                "of Perfection": "head/suffix/divine_hood_of_perfection.png",
-                "of Enlightenment": "head/suffix/divine_hood_of_enlightenment.png",
-                "of Rage": "head/suffix/divine_hood_of_rage.png",
-                "of the Fox": "head/suffix/divine_hood_of_the_fox.png"
-            },
-            "Ancient Helm": {
-                "of Detection": "head/suffix/ancient_helm_of_detection.png",
-                "of Enlightenment": "head/suffix/ancient_helm_of_enlightenment.png",
-                "of Skill": "head/suffix/ancient_helm_of_skill.png",
-                "of Giants": "head/suffix/ancient_helm_of_giants.png",
-                "of Brilliance": "head/suffix/ancient_helm_of_brilliance.png",
-                "of Anger": "head/suffix/ancient_helm_of_anger.png",
-                "of the Fox": "head/suffix/ancient_helm_of_the_fox.png",
-                "of Reflection": "head/suffix/ancient_helm_of_reflection.png",
-                "of Titans": "head/suffix/ancient_helm_of_titans.png",
-                "of Power": "head/suffix/ancient_helm_of_power.png",
-                "of Vitriol": "head/suffix/ancient_helm_of_vitriol.png",
-                "of Protection": "head/suffix/ancient_helm_of_protection.png",
-                "of Perfection": "head/suffix/ancient_helm_of_perfection.png",
-                "of the Twins": "head/suffix/ancient_helm_of_the_twins.png",
-                "of Fury": "head/suffix/ancient_helm_of_fury.png",
-                "of Rage": "head/suffix/ancient_helm_of_rage.png"
-            },
-            "Great Helm": {
-                "of Perfection": "head/suffix/great_helm_of_perfection.png",
-                "of Skill": "head/suffix/great_helm_of_skill.png",
-                "of Giants": "head/suffix/great_helm_of_giants.png",
-                "of Vitriol": "head/suffix/great_helm_of_vitriol.png",
-                "of Enlightenment": "head/suffix/great_helm_of_enlightenment.png",
-                "of Power": "head/suffix/great_helm_of_power.png",
-                "of Fury": "head/suffix/great_helm_of_fury.png",
-                "of Anger": "head/suffix/great_helm_of_anger.png",
-                "of Detection": "head/suffix/great_helm_of_detection.png",
-                "of the Twins": "head/suffix/great_helm_of_the_twins.png",
-                "of the Fox": "head/suffix/great_helm_of_the_fox.png",
-                "of Titans": "head/suffix/great_helm_of_titans.png",
-                "of Reflection": "head/suffix/great_helm_of_reflection.png",
-                "of Rage": "head/suffix/great_helm_of_rage.png",
-                "of Brilliance": "head/suffix/great_helm_of_brilliance.png",
-                "of Protection": "head/suffix/great_helm_of_protection.png"
-            },
-            "Silk Hood": {
-                "of Brilliance": "head/suffix/silk_hood_of_brilliance.png",
-                "of Skill": "head/suffix/silk_hood_of_skill.png",
-                "of Vitriol": "head/suffix/silk_hood_of_vitriol.png",
-                "of the Twins": "head/suffix/silk_hood_of_the_twins.png",
-                "of Rage": "head/suffix/silk_hood_of_rage.png",
-                "of Anger": "head/suffix/silk_hood_of_anger.png",
-                "of Titans": "head/suffix/silk_hood_of_titans.png",
-                "of Enlightenment": "head/suffix/silk_hood_of_enlightenment.png",
-                "of Detection": "head/suffix/silk_hood_of_detection.png",
-                "of Power": "head/suffix/silk_hood_of_power.png",
-                "of the Fox": "head/suffix/silk_hood_of_the_fox.png",
-                "of Protection": "head/suffix/silk_hood_of_protection.png",
-                "of Fury": "head/suffix/silk_hood_of_fury.png",
-                "of Reflection": "head/suffix/silk_hood_of_reflection.png",
-                "of Perfection": "head/suffix/silk_hood_of_perfection.png",
-                "of Giants": "head/suffix/silk_hood_of_giants.png"
-            },
-            "Dragon's Crown": {
-                "of Perfection": "head/suffix/dragons_crown_of_perfection.png",
-                "of Titans": "head/suffix/dragons_crown_of_titans.png",
-                "of Detection": "head/suffix/dragons_crown_of_detection.png",
-                "of Fury": "head/suffix/dragons_crown_of_fury.png",
-                "of Brilliance": "head/suffix/dragons_crown_of_brilliance.png",
-                "of Enlightenment": "head/suffix/dragons_crown_of_enlightenment.png",
-                "of Protection": "head/suffix/dragons_crown_of_protection.png",
-                "of Skill": "head/suffix/dragons_crown_of_skill.png",
-                "of Giants": "head/suffix/dragons_crown_of_giants.png",
-                "of Reflection": "head/suffix/dragons_crown_of_reflection.png",
-                "of the Fox": "head/suffix/dragons_crown_of_the_fox.png",
-                "of Anger": "head/suffix/dragons_crown_of_anger.png",
-                "of Rage": "head/suffix/dragons_crown_of_rage.png",
-                "of the Twins": "head/suffix/dragons_crown_of_the_twins.png",
-                "of Power": "head/suffix/dragons_crown_of_power.png",
-                "of Vitriol": "head/suffix/dragons_crown_of_vitriol.png"
-            },
-            "Hood": {
-                "of Skill": "head/suffix/hood_of_skill.png",
-                "of Perfection": "head/suffix/hood_of_perfection.png",
-                "of Titans": "head/suffix/hood_of_titans.png",
-                "of Giants": "head/suffix/hood_of_giants.png",
-                "of Brilliance": "head/suffix/hood_of_brilliance.png",
-                "of Fury": "head/suffix/hood_of_fury.png",
-                "of Protection": "head/suffix/hood_of_protection.png",
-                "of Reflection": "head/suffix/hood_of_reflection.png",
-                "of the Fox": "head/suffix/hood_of_the_fox.png",
-                "of Power": "head/suffix/hood_of_power.png",
-                "of Detection": "head/suffix/hood_of_detection.png",
-                "of Rage": "head/suffix/hood_of_rage.png",
-                "of the Twins": "head/suffix/hood_of_the_twins.png",
-                "of Vitriol": "head/suffix/hood_of_vitriol.png",
-                "of Anger": "head/suffix/hood_of_anger.png",
-                "of Enlightenment": "head/suffix/hood_of_enlightenment.png"
-            },
-            "Cap": {
-                "of Perfection": "head/suffix/cap_of_perfection.png",
-                "of Protection": "head/suffix/cap_of_protection.png",
-                "of Giants": "head/suffix/cap_of_giants.png",
-                "of Fury": "head/suffix/cap_of_fury.png",
-                "of the Fox": "head/suffix/cap_of_the_fox.png",
-                "of Rage": "head/suffix/cap_of_rage.png",
-                "of the Twins": "head/suffix/cap_of_the_twins.png",
-                "of Vitriol": "head/suffix/cap_of_vitriol.png",
-                "of Detection": "head/suffix/cap_of_detection.png",
-                "of Power": "head/suffix/cap_of_power.png",
-                "of Anger": "head/suffix/cap_of_anger.png",
-                "of Titans": "head/suffix/cap_of_titans.png",
-                "of Enlightenment": "head/suffix/cap_of_enlightenment.png",
-                "of Brilliance": "head/suffix/cap_of_brilliance.png",
-                "of Skill": "head/suffix/cap_of_skill.png",
-                "of Reflection": "head/suffix/cap_of_reflection.png"
-            },
-            "Full Helm": {
-                "of Anger": "head/suffix/full_helm_of_anger.png",
-                "of Skill": "head/suffix/full_helm_of_skill.png",
-                "of Detection": "head/suffix/full_helm_of_detection.png",
-                "of Reflection": "head/suffix/full_helm_of_reflection.png",
-                "of Brilliance": "head/suffix/full_helm_of_brilliance.png",
-                "of Rage": "head/suffix/full_helm_of_rage.png",
-                "of Vitriol": "head/suffix/full_helm_of_vitriol.png",
-                "of Titans": "head/suffix/full_helm_of_titans.png",
-                "of Power": "head/suffix/full_helm_of_power.png",
-                "of Enlightenment": "head/suffix/full_helm_of_enlightenment.png",
-                "of Giants": "head/suffix/full_helm_of_giants.png",
-                "of Protection": "head/suffix/full_helm_of_protection.png",
-                "of Fury": "head/suffix/full_helm_of_fury.png",
-                "of the Fox": "head/suffix/full_helm_of_the_fox.png",
-                "of Perfection": "head/suffix/full_helm_of_perfection.png",
-                "of the Twins": "head/suffix/full_helm_of_the_twins.png"
-            },
-            "Linen Hood": {
-                "of Fury": "head/suffix/linen_hood_of_fury.png",
-                "of the Twins": "head/suffix/linen_hood_of_the_twins.png",
-                "of Brilliance": "head/suffix/linen_hood_of_brilliance.png",
-                "of Perfection": "head/suffix/linen_hood_of_perfection.png",
-                "of Enlightenment": "head/suffix/linen_hood_of_enlightenment.png",
-                "of Giants": "head/suffix/linen_hood_of_giants.png",
-                "of Protection": "head/suffix/linen_hood_of_protection.png",
-                "of Reflection": "head/suffix/linen_hood_of_reflection.png",
-                "of Skill": "head/suffix/linen_hood_of_skill.png",
-                "of Power": "head/suffix/linen_hood_of_power.png",
-                "of Anger": "head/suffix/linen_hood_of_anger.png",
-                "of Titans": "head/suffix/linen_hood_of_titans.png",
-                "of Detection": "head/suffix/linen_hood_of_detection.png",
-                "of the Fox": "head/suffix/linen_hood_of_the_fox.png",
-                "of Rage": "head/suffix/linen_hood_of_rage.png",
-                "of Vitriol": "head/suffix/linen_hood_of_vitriol.png"
-            },
-            "Demon Crown": {
-                "of Skill": "head/suffix/demon_crown_of_skill.png",
-                "of the Fox": "head/suffix/demon_crown_of_the_fox.png",
-                "of Titans": "head/suffix/demon_crown_of_titans.png",
-                "of Protection": "head/suffix/demon_crown_of_protection.png",
-                "of Enlightenment": "head/suffix/demon_crown_of_enlightenment.png",
-                "of Detection": "head/suffix/demon_crown_of_detection.png",
-                "of the Twins": "head/suffix/demon_crown_of_the_twins.png",
-                "of Vitriol": "head/suffix/demon_crown_of_vitriol.png",
-                "of Rage": "head/suffix/demon_crown_of_rage.png",
-                "of Power": "head/suffix/demon_crown_of_power.png",
-                "of Anger": "head/suffix/demon_crown_of_anger.png",
-                "of Brilliance": "head/suffix/demon_crown_of_brilliance.png",
-                "of Fury": "head/suffix/demon_crown_of_fury.png",
-                "of Giants": "head/suffix/demon_crown_of_giants.png",
-                "of Perfection": "head/suffix/demon_crown_of_perfection.png",
-                "of Reflection": "head/suffix/demon_crown_of_reflection.png"
-            },
-            "Leather Cap": {
-                "of Perfection": "head/suffix/leather_cap_of_perfection.png",
-                "of Giants": "head/suffix/leather_cap_of_giants.png",
-                "of Vitriol": "head/suffix/leather_cap_of_vitriol.png",
-                "of Skill": "head/suffix/leather_cap_of_skill.png",
-                "of Rage": "head/suffix/leather_cap_of_rage.png",
-                "of Fury": "head/suffix/leather_cap_of_fury.png",
-                "of Detection": "head/suffix/leather_cap_of_detection.png",
-                "of Reflection": "head/suffix/leather_cap_of_reflection.png",
-                "of Protection": "head/suffix/leather_cap_of_protection.png",
-                "of Enlightenment": "head/suffix/leather_cap_of_enlightenment.png",
-                "of Power": "head/suffix/leather_cap_of_power.png",
-                "of Titans": "head/suffix/leather_cap_of_titans.png",
-                "of the Fox": "head/suffix/leather_cap_of_the_fox.png",
-                "of the Twins": "head/suffix/leather_cap_of_the_twins.png",
-                "of Anger": "head/suffix/leather_cap_of_anger.png",
-                "of Brilliance": "head/suffix/leather_cap_of_brilliance.png"
-            },
-            "Helm": {
-                "of Anger": "head/suffix/helm_of_anger.png",
-                "of Titans": "head/suffix/helm_of_titans.png",
-                "of Brilliance": "head/suffix/helm_of_brilliance.png",
-                "of Reflection": "head/suffix/helm_of_reflection.png",
-                "of Power": "head/suffix/helm_of_power.png",
-                "of Rage": "head/suffix/helm_of_rage.png",
-                "of Detection": "head/suffix/helm_of_detection.png",
-                "of Perfection": "head/suffix/helm_of_perfection.png",
-                "of Giants": "head/suffix/helm_of_giants.png",
-                "of the Fox": "head/suffix/helm_of_the_fox.png",
-                "of Vitriol": "head/suffix/helm_of_vitriol.png",
-                "of Enlightenment": "head/suffix/helm_of_enlightenment.png",
-                "of Skill": "head/suffix/helm_of_skill.png",
-                "of the Twins": "head/suffix/helm_of_the_twins.png",
-                "of Fury": "head/suffix/helm_of_fury.png",
-                "of Protection": "head/suffix/helm_of_protection.png"
-            },
-            "Ornate Helm": {
-                "of Giants": "head/suffix/ornate_helm_of_giants.png",
-                "of Protection": "head/suffix/ornate_helm_of_protection.png",
-                "of Skill": "head/suffix/ornate_helm_of_skill.png",
-                "of Titans": "head/suffix/ornate_helm_of_titans.png",
-                "of Rage": "head/suffix/ornate_helm_of_rage.png",
-                "of Vitriol": "head/suffix/ornate_helm_of_vitriol.png",
-                "of Fury": "head/suffix/ornate_helm_of_fury.png",
-                "of Detection": "head/suffix/ornate_helm_of_detection.png",
-                "of Power": "head/suffix/ornate_helm_of_power.png",
-                "of Brilliance": "head/suffix/ornate_helm_of_brilliance.png",
-                "of Anger": "head/suffix/ornate_helm_of_anger.png",
-                "of the Twins": "head/suffix/ornate_helm_of_the_twins.png",
-                "of Enlightenment": "head/suffix/ornate_helm_of_enlightenment.png",
-                "of Reflection": "head/suffix/ornate_helm_of_reflection.png",
-                "of the Fox": "head/suffix/ornate_helm_of_the_fox.png",
-                "of Perfection": "head/suffix/ornate_helm_of_perfection.png"
-            },
-            "Crown": {
-                "of Reflection": "head/suffix/crown_of_reflection.png",
-                "of Fury": "head/suffix/crown_of_fury.png",
-                "of Vitriol": "head/suffix/crown_of_vitriol.png",
-                "of Enlightenment": "head/suffix/crown_of_enlightenment.png",
-                "of Rage": "head/suffix/crown_of_rage.png",
-                "of Detection": "head/suffix/crown_of_detection.png",
-                "of Perfection": "head/suffix/crown_of_perfection.png",
-                "of Anger": "head/suffix/crown_of_anger.png",
-                "of the Twins": "head/suffix/crown_of_the_twins.png",
-                "of Skill": "head/suffix/crown_of_skill.png",
-                "of Titans": "head/suffix/crown_of_titans.png",
-                "of the Fox": "head/suffix/crown_of_the_fox.png",
-                "of Giants": "head/suffix/crown_of_giants.png",
-                "of Protection": "head/suffix/crown_of_protection.png",
-                "of Brilliance": "head/suffix/crown_of_brilliance.png",
-                "of Power": "head/suffix/crown_of_power.png"
-            },
-            "War Cap": {
-                "of the Twins": "head/suffix/war_cap_of_the_twins.png",
-                "of Perfection": "head/suffix/war_cap_of_perfection.png",
-                "of Brilliance": "head/suffix/war_cap_of_brilliance.png",
-                "of Fury": "head/suffix/war_cap_of_fury.png",
-                "of Giants": "head/suffix/war_cap_of_giants.png",
-                "of Vitriol": "head/suffix/war_cap_of_vitriol.png",
-                "of Enlightenment": "head/suffix/war_cap_of_enlightenment.png",
-                "of Protection": "head/suffix/war_cap_of_protection.png",
-                "of Rage": "head/suffix/war_cap_of_rage.png",
-                "of Anger": "head/suffix/war_cap_of_anger.png",
-                "of Reflection": "head/suffix/war_cap_of_reflection.png",
-                "of Detection": "head/suffix/war_cap_of_detection.png",
-                "of Skill": "head/suffix/war_cap_of_skill.png",
-                "of Titans": "head/suffix/war_cap_of_titans.png",
-                "of Power": "head/suffix/war_cap_of_power.png",
-                "of the Fox": "head/suffix/war_cap_of_the_fox.png"
-            }
-        },
-        "plus_one": "head/plus_one/plus_one.png",
-        "name": {
-            "Divine Hood": "head/name/divine_hood.png",
-            "Ancient Helm": "head/name/ancient_helm.png",
-            "Great Helm": "head/name/great_helm.png",
-            "Silk Hood": "head/name/silk_hood.png",
-            "Dragon's Crown": "head/name/dragons_crown.png",
-            "Hood": "head/name/hood.png",
-            "Cap": "head/name/cap.png",
-            "Full Helm": "head/name/full_helm.png",
-            "Linen Hood": "head/name/linen_hood.png",
-            "Demon Crown": "head/name/demon_crown.png",
-            "Leather Cap": "head/name/leather_cap.png",
-            "Helm": "head/name/helm.png",
-            "Ornate Helm": "head/name/ornate_helm.png",
-            "Crown": "head/name/crown.png",
-            "War Cap": "head/name/war_cap.png"
-        }
-    },
-    "waist": {
-        "prefix": "waist/prefix/prefix.png",
-        "suffix": {
-            "Hard Leather Belt": {
-                "of Vitriol": "waist/suffix/hard_leather_belt_of_vitriol.png",
-                "of Giants": "waist/suffix/hard_leather_belt_of_giants.png",
-                "of Rage": "waist/suffix/hard_leather_belt_of_rage.png",
-                "of Skill": "waist/suffix/hard_leather_belt_of_skill.png",
-                "of Anger": "waist/suffix/hard_leather_belt_of_anger.png",
-                "of Enlightenment": "waist/suffix/hard_leather_belt_of_enlightenment.png",
-                "of Perfection": "waist/suffix/hard_leather_belt_of_perfection.png",
-                "of Reflection": "waist/suffix/hard_leather_belt_of_reflection.png",
-                "of Detection": "waist/suffix/hard_leather_belt_of_detection.png",
-                "of Power": "waist/suffix/hard_leather_belt_of_power.png",
-                "of the Twins": "waist/suffix/hard_leather_belt_of_the_twins.png",
-                "of Protection": "waist/suffix/hard_leather_belt_of_protection.png",
-                "of Fury": "waist/suffix/hard_leather_belt_of_fury.png",
-                "of Brilliance": "waist/suffix/hard_leather_belt_of_brilliance.png",
-                "of Titans": "waist/suffix/hard_leather_belt_of_titans.png",
-                "of the Fox": "waist/suffix/hard_leather_belt_of_the_fox.png"
-            },
-            "Plated Belt": {
-                "of Enlightenment": "waist/suffix/plated_belt_of_enlightenment.png",
-                "of Detection": "waist/suffix/plated_belt_of_detection.png",
-                "of Titans": "waist/suffix/plated_belt_of_titans.png",
-                "of Protection": "waist/suffix/plated_belt_of_protection.png",
-                "of the Twins": "waist/suffix/plated_belt_of_the_twins.png",
-                "of Perfection": "waist/suffix/plated_belt_of_perfection.png",
-                "of Giants": "waist/suffix/plated_belt_of_giants.png",
-                "of Reflection": "waist/suffix/plated_belt_of_reflection.png",
-                "of Vitriol": "waist/suffix/plated_belt_of_vitriol.png",
-                "of Brilliance": "waist/suffix/plated_belt_of_brilliance.png",
-                "of Rage": "waist/suffix/plated_belt_of_rage.png",
-                "of Fury": "waist/suffix/plated_belt_of_fury.png",
-                "of the Fox": "waist/suffix/plated_belt_of_the_fox.png",
-                "of Power": "waist/suffix/plated_belt_of_power.png",
-                "of Anger": "waist/suffix/plated_belt_of_anger.png",
-                "of Skill": "waist/suffix/plated_belt_of_skill.png"
-            },
-            "Ornate Belt": {
-                "of Detection": "waist/suffix/ornate_belt_of_detection.png",
-                "of Titans": "waist/suffix/ornate_belt_of_titans.png",
-                "of Brilliance": "waist/suffix/ornate_belt_of_brilliance.png",
-                "of Giants": "waist/suffix/ornate_belt_of_giants.png",
-                "of Rage": "waist/suffix/ornate_belt_of_rage.png",
-                "of Fury": "waist/suffix/ornate_belt_of_fury.png",
-                "of Skill": "waist/suffix/ornate_belt_of_skill.png",
-                "of Anger": "waist/suffix/ornate_belt_of_anger.png",
-                "of Enlightenment": "waist/suffix/ornate_belt_of_enlightenment.png",
-                "of Reflection": "waist/suffix/ornate_belt_of_reflection.png",
-                "of the Fox": "waist/suffix/ornate_belt_of_the_fox.png",
-                "of Perfection": "waist/suffix/ornate_belt_of_perfection.png",
-                "of the Twins": "waist/suffix/ornate_belt_of_the_twins.png",
-                "of Power": "waist/suffix/ornate_belt_of_power.png",
-                "of Protection": "waist/suffix/ornate_belt_of_protection.png",
-                "of Vitriol": "waist/suffix/ornate_belt_of_vitriol.png"
-            },
-            "Heavy Belt": {
-                "of the Fox": "waist/suffix/heavy_belt_of_the_fox.png",
-                "of Rage": "waist/suffix/heavy_belt_of_rage.png",
-                "of Detection": "waist/suffix/heavy_belt_of_detection.png",
-                "of the Twins": "waist/suffix/heavy_belt_of_the_twins.png",
-                "of Perfection": "waist/suffix/heavy_belt_of_perfection.png",
-                "of Anger": "waist/suffix/heavy_belt_of_anger.png",
-                "of Power": "waist/suffix/heavy_belt_of_power.png",
-                "of Titans": "waist/suffix/heavy_belt_of_titans.png",
-                "of Skill": "waist/suffix/heavy_belt_of_skill.png",
-                "of Reflection": "waist/suffix/heavy_belt_of_reflection.png",
-                "of Fury": "waist/suffix/heavy_belt_of_fury.png",
-                "of Brilliance": "waist/suffix/heavy_belt_of_brilliance.png",
-                "of Vitriol": "waist/suffix/heavy_belt_of_vitriol.png",
-                "of Giants": "waist/suffix/heavy_belt_of_giants.png",
-                "of Enlightenment": "waist/suffix/heavy_belt_of_enlightenment.png",
-                "of Protection": "waist/suffix/heavy_belt_of_protection.png"
-            },
-            "Sash": {
-                "of Fury": "waist/suffix/sash_of_fury.png",
-                "of Giants": "waist/suffix/sash_of_giants.png",
-                "of Detection": "waist/suffix/sash_of_detection.png",
-                "of Anger": "waist/suffix/sash_of_anger.png",
-                "of the Twins": "waist/suffix/sash_of_the_twins.png",
-                "of Rage": "waist/suffix/sash_of_rage.png",
-                "of Skill": "waist/suffix/sash_of_skill.png",
-                "of Vitriol": "waist/suffix/sash_of_vitriol.png",
-                "of Protection": "waist/suffix/sash_of_protection.png",
-                "of the Fox": "waist/suffix/sash_of_the_fox.png",
-                "of Enlightenment": "waist/suffix/sash_of_enlightenment.png",
-                "of Power": "waist/suffix/sash_of_power.png",
-                "of Perfection": "waist/suffix/sash_of_perfection.png",
-                "of Titans": "waist/suffix/sash_of_titans.png",
-                "of Reflection": "waist/suffix/sash_of_reflection.png",
-                "of Brilliance": "waist/suffix/sash_of_brilliance.png"
-            },
-            "Leather Belt": {
-                "of the Fox": "waist/suffix/leather_belt_of_the_fox.png",
-                "of Enlightenment": "waist/suffix/leather_belt_of_enlightenment.png",
-                "of Reflection": "waist/suffix/leather_belt_of_reflection.png",
-                "of Rage": "waist/suffix/leather_belt_of_rage.png",
-                "of Perfection": "waist/suffix/leather_belt_of_perfection.png",
-                "of Anger": "waist/suffix/leather_belt_of_anger.png",
-                "of Giants": "waist/suffix/leather_belt_of_giants.png",
-                "of Protection": "waist/suffix/leather_belt_of_protection.png",
-                "of Skill": "waist/suffix/leather_belt_of_skill.png",
-                "of Brilliance": "waist/suffix/leather_belt_of_brilliance.png",
-                "of Vitriol": "waist/suffix/leather_belt_of_vitriol.png",
-                "of Power": "waist/suffix/leather_belt_of_power.png",
-                "of Detection": "waist/suffix/leather_belt_of_detection.png",
-                "of Fury": "waist/suffix/leather_belt_of_fury.png",
-                "of Titans": "waist/suffix/leather_belt_of_titans.png",
-                "of the Twins": "waist/suffix/leather_belt_of_the_twins.png"
-            },
-            "Linen Sash": {
-                "of Reflection": "waist/suffix/linen_sash_of_reflection.png",
-                "of the Twins": "waist/suffix/linen_sash_of_the_twins.png",
-                "of Protection": "waist/suffix/linen_sash_of_protection.png",
-                "of Perfection": "waist/suffix/linen_sash_of_perfection.png",
-                "of Brilliance": "waist/suffix/linen_sash_of_brilliance.png",
-                "of Enlightenment": "waist/suffix/linen_sash_of_enlightenment.png",
-                "of Detection": "waist/suffix/linen_sash_of_detection.png",
-                "of Anger": "waist/suffix/linen_sash_of_anger.png",
-                "of the Fox": "waist/suffix/linen_sash_of_the_fox.png",
-                "of Vitriol": "waist/suffix/linen_sash_of_vitriol.png",
-                "of Skill": "waist/suffix/linen_sash_of_skill.png",
-                "of Fury": "waist/suffix/linen_sash_of_fury.png",
-                "of Power": "waist/suffix/linen_sash_of_power.png",
-                "of Titans": "waist/suffix/linen_sash_of_titans.png",
-                "of Giants": "waist/suffix/linen_sash_of_giants.png",
-                "of Rage": "waist/suffix/linen_sash_of_rage.png"
-            },
-            "War Belt": {
-                "of Perfection": "waist/suffix/war_belt_of_perfection.png",
-                "of Power": "waist/suffix/war_belt_of_power.png",
-                "of Skill": "waist/suffix/war_belt_of_skill.png",
-                "of Enlightenment": "waist/suffix/war_belt_of_enlightenment.png",
-                "of Protection": "waist/suffix/war_belt_of_protection.png",
-                "of Titans": "waist/suffix/war_belt_of_titans.png",
-                "of Reflection": "waist/suffix/war_belt_of_reflection.png",
-                "of the Twins": "waist/suffix/war_belt_of_the_twins.png",
-                "of Giants": "waist/suffix/war_belt_of_giants.png",
-                "of Anger": "waist/suffix/war_belt_of_anger.png",
-                "of Vitriol": "waist/suffix/war_belt_of_vitriol.png",
-                "of Fury": "waist/suffix/war_belt_of_fury.png",
-                "of Detection": "waist/suffix/war_belt_of_detection.png",
-                "of Brilliance": "waist/suffix/war_belt_of_brilliance.png",
-                "of the Fox": "waist/suffix/war_belt_of_the_fox.png",
-                "of Rage": "waist/suffix/war_belt_of_rage.png"
-            },
-            "Mesh Belt": {
-                "of Skill": "waist/suffix/mesh_belt_of_skill.png",
-                "of Rage": "waist/suffix/mesh_belt_of_rage.png",
-                "of Vitriol": "waist/suffix/mesh_belt_of_vitriol.png",
-                "of Titans": "waist/suffix/mesh_belt_of_titans.png",
-                "of Fury": "waist/suffix/mesh_belt_of_fury.png",
-                "of Enlightenment": "waist/suffix/mesh_belt_of_enlightenment.png",
-                "of Anger": "waist/suffix/mesh_belt_of_anger.png",
-                "of Giants": "waist/suffix/mesh_belt_of_giants.png",
-                "of the Twins": "waist/suffix/mesh_belt_of_the_twins.png",
-                "of Detection": "waist/suffix/mesh_belt_of_detection.png",
-                "of Reflection": "waist/suffix/mesh_belt_of_reflection.png",
-                "of Perfection": "waist/suffix/mesh_belt_of_perfection.png",
-                "of the Fox": "waist/suffix/mesh_belt_of_the_fox.png",
-                "of Protection": "waist/suffix/mesh_belt_of_protection.png",
-                "of Brilliance": "waist/suffix/mesh_belt_of_brilliance.png",
-                "of Power": "waist/suffix/mesh_belt_of_power.png"
-            },
-            "Studded Leather Belt": {
-                "of Brilliance": "waist/suffix/studded_leather_belt_of_brilliance.png",
-                "of Vitriol": "waist/suffix/studded_leather_belt_of_vitriol.png",
-                "of Rage": "waist/suffix/studded_leather_belt_of_rage.png",
-                "of Protection": "waist/suffix/studded_leather_belt_of_protection.png",
-                "of Perfection": "waist/suffix/studded_leather_belt_of_perfection.png",
-                "of Reflection": "waist/suffix/studded_leather_belt_of_reflection.png",
-                "of Skill": "waist/suffix/studded_leather_belt_of_skill.png",
-                "of Fury": "waist/suffix/studded_leather_belt_of_fury.png",
-                "of Giants": "waist/suffix/studded_leather_belt_of_giants.png",
-                "of Enlightenment": "waist/suffix/studded_leather_belt_of_enlightenment.png",
-                "of Power": "waist/suffix/studded_leather_belt_of_power.png",
-                "of Titans": "waist/suffix/studded_leather_belt_of_titans.png",
-                "of the Fox": "waist/suffix/studded_leather_belt_of_the_fox.png",
-                "of Detection": "waist/suffix/studded_leather_belt_of_detection.png",
-                "of the Twins": "waist/suffix/studded_leather_belt_of_the_twins.png",
-                "of Anger": "waist/suffix/studded_leather_belt_of_anger.png"
-            },
-            "Silk Sash": {
-                "of Skill": "waist/suffix/silk_sash_of_skill.png",
-                "of the Fox": "waist/suffix/silk_sash_of_the_fox.png",
-                "of Rage": "waist/suffix/silk_sash_of_rage.png",
-                "of Fury": "waist/suffix/silk_sash_of_fury.png",
-                "of Brilliance": "waist/suffix/silk_sash_of_brilliance.png",
-                "of Power": "waist/suffix/silk_sash_of_power.png",
-                "of Protection": "waist/suffix/silk_sash_of_protection.png",
-                "of Anger": "waist/suffix/silk_sash_of_anger.png",
-                "of Enlightenment": "waist/suffix/silk_sash_of_enlightenment.png",
-                "of Giants": "waist/suffix/silk_sash_of_giants.png",
-                "of Detection": "waist/suffix/silk_sash_of_detection.png",
-                "of Titans": "waist/suffix/silk_sash_of_titans.png",
-                "of Vitriol": "waist/suffix/silk_sash_of_vitriol.png",
-                "of Perfection": "waist/suffix/silk_sash_of_perfection.png",
-                "of the Twins": "waist/suffix/silk_sash_of_the_twins.png",
-                "of Reflection": "waist/suffix/silk_sash_of_reflection.png"
-            },
-            "Brightsilk Sash": {
-                "of Fury": "waist/suffix/brightsilk_sash_of_fury.png",
-                "of Enlightenment": "waist/suffix/brightsilk_sash_of_enlightenment.png",
-                "of Vitriol": "waist/suffix/brightsilk_sash_of_vitriol.png",
-                "of Anger": "waist/suffix/brightsilk_sash_of_anger.png",
-                "of Perfection": "waist/suffix/brightsilk_sash_of_perfection.png",
-                "of Power": "waist/suffix/brightsilk_sash_of_power.png",
-                "of Rage": "waist/suffix/brightsilk_sash_of_rage.png",
-                "of Reflection": "waist/suffix/brightsilk_sash_of_reflection.png",
-                "of Protection": "waist/suffix/brightsilk_sash_of_protection.png",
-                "of Titans": "waist/suffix/brightsilk_sash_of_titans.png",
-                "of the Twins": "waist/suffix/brightsilk_sash_of_the_twins.png",
-                "of Detection": "waist/suffix/brightsilk_sash_of_detection.png",
-                "of Skill": "waist/suffix/brightsilk_sash_of_skill.png",
-                "of the Fox": "waist/suffix/brightsilk_sash_of_the_fox.png",
-                "of Giants": "waist/suffix/brightsilk_sash_of_giants.png",
-                "of Brilliance": "waist/suffix/brightsilk_sash_of_brilliance.png"
-            },
-            "Dragonskin Belt": {
-                "of Giants": "waist/suffix/dragonskin_belt_of_giants.png",
-                "of Reflection": "waist/suffix/dragonskin_belt_of_reflection.png",
-                "of Titans": "waist/suffix/dragonskin_belt_of_titans.png",
-                "of Enlightenment": "waist/suffix/dragonskin_belt_of_enlightenment.png",
-                "of Power": "waist/suffix/dragonskin_belt_of_power.png",
-                "of Anger": "waist/suffix/dragonskin_belt_of_anger.png",
-                "of Detection": "waist/suffix/dragonskin_belt_of_detection.png",
-                "of Fury": "waist/suffix/dragonskin_belt_of_fury.png",
-                "of the Twins": "waist/suffix/dragonskin_belt_of_the_twins.png",
-                "of Rage": "waist/suffix/dragonskin_belt_of_rage.png",
-                "of Skill": "waist/suffix/dragonskin_belt_of_skill.png",
-                "of Brilliance": "waist/suffix/dragonskin_belt_of_brilliance.png",
-                "of Protection": "waist/suffix/dragonskin_belt_of_protection.png",
-                "of Perfection": "waist/suffix/dragonskin_belt_of_perfection.png",
-                "of the Fox": "waist/suffix/dragonskin_belt_of_the_fox.png",
-                "of Vitriol": "waist/suffix/dragonskin_belt_of_vitriol.png"
-            },
-            "Wool Sash": {
-                "of Detection": "waist/suffix/wool_sash_of_detection.png",
-                "of Protection": "waist/suffix/wool_sash_of_protection.png",
-                "of Vitriol": "waist/suffix/wool_sash_of_vitriol.png",
-                "of the Twins": "waist/suffix/wool_sash_of_the_twins.png",
-                "of Anger": "waist/suffix/wool_sash_of_anger.png",
-                "of Brilliance": "waist/suffix/wool_sash_of_brilliance.png",
-                "of Fury": "waist/suffix/wool_sash_of_fury.png",
-                "of Giants": "waist/suffix/wool_sash_of_giants.png",
-                "of Perfection": "waist/suffix/wool_sash_of_perfection.png",
-                "of Titans": "waist/suffix/wool_sash_of_titans.png",
-                "of Rage": "waist/suffix/wool_sash_of_rage.png",
-                "of Enlightenment": "waist/suffix/wool_sash_of_enlightenment.png",
-                "of Reflection": "waist/suffix/wool_sash_of_reflection.png",
-                "of Power": "waist/suffix/wool_sash_of_power.png",
-                "of the Fox": "waist/suffix/wool_sash_of_the_fox.png",
-                "of Skill": "waist/suffix/wool_sash_of_skill.png"
-            },
-            "Demonhide Belt": {
-                "of Rage": "waist/suffix/demonhide_belt_of_rage.png",
-                "of Power": "waist/suffix/demonhide_belt_of_power.png",
-                "of the Twins": "waist/suffix/demonhide_belt_of_the_twins.png",
-                "of Reflection": "waist/suffix/demonhide_belt_of_reflection.png",
-                "of Perfection": "waist/suffix/demonhide_belt_of_perfection.png",
-                "of Enlightenment": "waist/suffix/demonhide_belt_of_enlightenment.png",
-                "of Protection": "waist/suffix/demonhide_belt_of_protection.png",
-                "of Skill": "waist/suffix/demonhide_belt_of_skill.png",
-                "of Fury": "waist/suffix/demonhide_belt_of_fury.png",
-                "of Giants": "waist/suffix/demonhide_belt_of_giants.png",
-                "of Anger": "waist/suffix/demonhide_belt_of_anger.png",
-                "of Titans": "waist/suffix/demonhide_belt_of_titans.png",
-                "of the Fox": "waist/suffix/demonhide_belt_of_the_fox.png",
-                "of Detection": "waist/suffix/demonhide_belt_of_detection.png",
-                "of Vitriol": "waist/suffix/demonhide_belt_of_vitriol.png",
-                "of Brilliance": "waist/suffix/demonhide_belt_of_brilliance.png"
-            }
-        },
-        "plus_one": "waist/plus_one/plus_one.png",
-        "name": {
-            "Hard Leather Belt": "waist/name/hard_leather_belt.png",
-            "Plated Belt": "waist/name/plated_belt.png",
-            "Ornate Belt": "waist/name/ornate_belt.png",
-            "Heavy Belt": "waist/name/heavy_belt.png",
-            "Sash": "waist/name/sash.png",
-            "Leather Belt": "waist/name/leather_belt.png",
-            "Linen Sash": "waist/name/linen_sash.png",
-            "War Belt": "waist/name/war_belt.png",
-            "Mesh Belt": "waist/name/mesh_belt.png",
-            "Studded Leather Belt": "waist/name/studded_leather_belt.png",
-            "Silk Sash": "waist/name/silk_sash.png",
-            "Brightsilk Sash": "waist/name/brightsilk_sash.png",
-            "Dragonskin Belt": "waist/name/dragonskin_belt.png",
-            "Wool Sash": "waist/name/wool_sash.png",
-            "Demonhide Belt": "waist/name/demonhide_belt.png"
-        }
-    },
-    "foot": {
-        "prefix": "foot/prefix/prefix.png",
-        "suffix": {
-            "Ornate Greaves": {
-                "of Skill": "foot/suffix/ornate_greaves_of_skill.png",
-                "of Anger": "foot/suffix/ornate_greaves_of_anger.png",
-                "of Perfection": "foot/suffix/ornate_greaves_of_perfection.png",
-                "of Reflection": "foot/suffix/ornate_greaves_of_reflection.png",
-                "of Protection": "foot/suffix/ornate_greaves_of_protection.png",
-                "of Power": "foot/suffix/ornate_greaves_of_power.png",
-                "of the Fox": "foot/suffix/ornate_greaves_of_the_fox.png",
-                "of the Twins": "foot/suffix/ornate_greaves_of_the_twins.png",
-                "of Brilliance": "foot/suffix/ornate_greaves_of_brilliance.png",
-                "of Enlightenment": "foot/suffix/ornate_greaves_of_enlightenment.png",
-                "of Vitriol": "foot/suffix/ornate_greaves_of_vitriol.png",
-                "of Fury": "foot/suffix/ornate_greaves_of_fury.png",
-                "of Rage": "foot/suffix/ornate_greaves_of_rage.png",
-                "of Giants": "foot/suffix/ornate_greaves_of_giants.png",
-                "of Titans": "foot/suffix/ornate_greaves_of_titans.png",
-                "of Detection": "foot/suffix/ornate_greaves_of_detection.png"
-            },
-            "Demonhide Boots": {
-                "of Brilliance": "foot/suffix/demonhide_boots_of_brilliance.png",
-                "of Perfection": "foot/suffix/demonhide_boots_of_perfection.png",
-                "of the Fox": "foot/suffix/demonhide_boots_of_the_fox.png",
-                "of Detection": "foot/suffix/demonhide_boots_of_detection.png",
-                "of Enlightenment": "foot/suffix/demonhide_boots_of_enlightenment.png",
-                "of the Twins": "foot/suffix/demonhide_boots_of_the_twins.png",
-                "of Rage": "foot/suffix/demonhide_boots_of_rage.png",
-                "of Titans": "foot/suffix/demonhide_boots_of_titans.png",
-                "of Fury": "foot/suffix/demonhide_boots_of_fury.png",
-                "of Reflection": "foot/suffix/demonhide_boots_of_reflection.png",
-                "of Protection": "foot/suffix/demonhide_boots_of_protection.png",
-                "of Skill": "foot/suffix/demonhide_boots_of_skill.png",
-                "of Vitriol": "foot/suffix/demonhide_boots_of_vitriol.png",
-                "of Anger": "foot/suffix/demonhide_boots_of_anger.png",
-                "of Giants": "foot/suffix/demonhide_boots_of_giants.png",
-                "of Power": "foot/suffix/demonhide_boots_of_power.png"
-            },
-            "Holy Greaves": {
-                "of Giants": "foot/suffix/holy_greaves_of_giants.png",
-                "of Enlightenment": "foot/suffix/holy_greaves_of_enlightenment.png",
-                "of Brilliance": "foot/suffix/holy_greaves_of_brilliance.png",
-                "of the Fox": "foot/suffix/holy_greaves_of_the_fox.png",
-                "of Vitriol": "foot/suffix/holy_greaves_of_vitriol.png",
-                "of Titans": "foot/suffix/holy_greaves_of_titans.png",
-                "of Fury": "foot/suffix/holy_greaves_of_fury.png",
-                "of Perfection": "foot/suffix/holy_greaves_of_perfection.png",
-                "of Skill": "foot/suffix/holy_greaves_of_skill.png",
-                "of Protection": "foot/suffix/holy_greaves_of_protection.png",
-                "of Reflection": "foot/suffix/holy_greaves_of_reflection.png",
-                "of Power": "foot/suffix/holy_greaves_of_power.png",
-                "of the Twins": "foot/suffix/holy_greaves_of_the_twins.png",
-                "of Anger": "foot/suffix/holy_greaves_of_anger.png",
-                "of Rage": "foot/suffix/holy_greaves_of_rage.png",
-                "of Detection": "foot/suffix/holy_greaves_of_detection.png"
-            },
-            "Dragonskin Boots": {
-                "of Fury": "foot/suffix/dragonskin_boots_of_fury.png",
-                "of Detection": "foot/suffix/dragonskin_boots_of_detection.png",
-                "of Titans": "foot/suffix/dragonskin_boots_of_titans.png",
-                "of Giants": "foot/suffix/dragonskin_boots_of_giants.png",
-                "of Perfection": "foot/suffix/dragonskin_boots_of_perfection.png",
-                "of Anger": "foot/suffix/dragonskin_boots_of_anger.png",
-                "of Vitriol": "foot/suffix/dragonskin_boots_of_vitriol.png",
-                "of the Fox": "foot/suffix/dragonskin_boots_of_the_fox.png",
-                "of Skill": "foot/suffix/dragonskin_boots_of_skill.png",
-                "of Brilliance": "foot/suffix/dragonskin_boots_of_brilliance.png",
-                "of the Twins": "foot/suffix/dragonskin_boots_of_the_twins.png",
-                "of Reflection": "foot/suffix/dragonskin_boots_of_reflection.png",
-                "of Rage": "foot/suffix/dragonskin_boots_of_rage.png",
-                "of Protection": "foot/suffix/dragonskin_boots_of_protection.png",
-                "of Power": "foot/suffix/dragonskin_boots_of_power.png",
-                "of Enlightenment": "foot/suffix/dragonskin_boots_of_enlightenment.png"
-            },
-            "Greaves": {
-                "of Enlightenment": "foot/suffix/greaves_of_enlightenment.png",
-                "of Reflection": "foot/suffix/greaves_of_reflection.png",
-                "of Giants": "foot/suffix/greaves_of_giants.png",
-                "of Perfection": "foot/suffix/greaves_of_perfection.png",
-                "of the Fox": "foot/suffix/greaves_of_the_fox.png",
-                "of the Twins": "foot/suffix/greaves_of_the_twins.png",
-                "of Vitriol": "foot/suffix/greaves_of_vitriol.png",
-                "of Skill": "foot/suffix/greaves_of_skill.png",
-                "of Power": "foot/suffix/greaves_of_power.png",
-                "of Rage": "foot/suffix/greaves_of_rage.png",
-                "of Protection": "foot/suffix/greaves_of_protection.png",
-                "of Titans": "foot/suffix/greaves_of_titans.png",
-                "of Fury": "foot/suffix/greaves_of_fury.png",
-                "of Anger": "foot/suffix/greaves_of_anger.png",
-                "of Brilliance": "foot/suffix/greaves_of_brilliance.png",
-                "of Detection": "foot/suffix/greaves_of_detection.png"
-            },
-            "Hard Leather Boots": {
-                "of Reflection": "foot/suffix/hard_leather_boots_of_reflection.png",
-                "of Vitriol": "foot/suffix/hard_leather_boots_of_vitriol.png",
-                "of Rage": "foot/suffix/hard_leather_boots_of_rage.png",
-                "of Fury": "foot/suffix/hard_leather_boots_of_fury.png",
-                "of Titans": "foot/suffix/hard_leather_boots_of_titans.png",
-                "of the Fox": "foot/suffix/hard_leather_boots_of_the_fox.png",
-                "of Protection": "foot/suffix/hard_leather_boots_of_protection.png",
-                "of Anger": "foot/suffix/hard_leather_boots_of_anger.png",
-                "of Skill": "foot/suffix/hard_leather_boots_of_skill.png",
-                "of Giants": "foot/suffix/hard_leather_boots_of_giants.png",
-                "of the Twins": "foot/suffix/hard_leather_boots_of_the_twins.png",
-                "of Perfection": "foot/suffix/hard_leather_boots_of_perfection.png",
-                "of Detection": "foot/suffix/hard_leather_boots_of_detection.png",
-                "of Power": "foot/suffix/hard_leather_boots_of_power.png",
-                "of Brilliance": "foot/suffix/hard_leather_boots_of_brilliance.png",
-                "of Enlightenment": "foot/suffix/hard_leather_boots_of_enlightenment.png"
-            },
-            "Wool Shoes": {
-                "of the Twins": "foot/suffix/wool_shoes_of_the_twins.png",
-                "of Brilliance": "foot/suffix/wool_shoes_of_brilliance.png",
-                "of Reflection": "foot/suffix/wool_shoes_of_reflection.png",
-                "of Anger": "foot/suffix/wool_shoes_of_anger.png",
-                "of Skill": "foot/suffix/wool_shoes_of_skill.png",
-                "of Detection": "foot/suffix/wool_shoes_of_detection.png",
-                "of Enlightenment": "foot/suffix/wool_shoes_of_enlightenment.png",
-                "of Vitriol": "foot/suffix/wool_shoes_of_vitriol.png",
-                "of the Fox": "foot/suffix/wool_shoes_of_the_fox.png",
-                "of Perfection": "foot/suffix/wool_shoes_of_perfection.png",
-                "of Fury": "foot/suffix/wool_shoes_of_fury.png",
-                "of Power": "foot/suffix/wool_shoes_of_power.png",
-                "of Giants": "foot/suffix/wool_shoes_of_giants.png",
-                "of Rage": "foot/suffix/wool_shoes_of_rage.png",
-                "of Protection": "foot/suffix/wool_shoes_of_protection.png",
-                "of Titans": "foot/suffix/wool_shoes_of_titans.png"
-            },
-            "Studded Leather Boots": {
-                "of the Twins": "foot/suffix/studded_leather_boots_of_the_twins.png",
-                "of Perfection": "foot/suffix/studded_leather_boots_of_perfection.png",
-                "of Enlightenment": "foot/suffix/studded_leather_boots_of_enlightenment.png",
-                "of Vitriol": "foot/suffix/studded_leather_boots_of_vitriol.png",
-                "of Power": "foot/suffix/studded_leather_boots_of_power.png",
-                "of the Fox": "foot/suffix/studded_leather_boots_of_the_fox.png",
-                "of Reflection": "foot/suffix/studded_leather_boots_of_reflection.png",
-                "of Protection": "foot/suffix/studded_leather_boots_of_protection.png",
-                "of Detection": "foot/suffix/studded_leather_boots_of_detection.png",
-                "of Anger": "foot/suffix/studded_leather_boots_of_anger.png",
-                "of Rage": "foot/suffix/studded_leather_boots_of_rage.png",
-                "of Giants": "foot/suffix/studded_leather_boots_of_giants.png",
-                "of Fury": "foot/suffix/studded_leather_boots_of_fury.png",
-                "of Skill": "foot/suffix/studded_leather_boots_of_skill.png",
-                "of Brilliance": "foot/suffix/studded_leather_boots_of_brilliance.png",
-                "of Titans": "foot/suffix/studded_leather_boots_of_titans.png"
-            },
-            "Heavy Boots": {
-                "of Protection": "foot/suffix/heavy_boots_of_protection.png",
-                "of Detection": "foot/suffix/heavy_boots_of_detection.png",
-                "of Anger": "foot/suffix/heavy_boots_of_anger.png",
-                "of Power": "foot/suffix/heavy_boots_of_power.png",
-                "of the Twins": "foot/suffix/heavy_boots_of_the_twins.png",
-                "of Perfection": "foot/suffix/heavy_boots_of_perfection.png",
-                "of Enlightenment": "foot/suffix/heavy_boots_of_enlightenment.png",
-                "of Rage": "foot/suffix/heavy_boots_of_rage.png",
-                "of the Fox": "foot/suffix/heavy_boots_of_the_fox.png",
-                "of Brilliance": "foot/suffix/heavy_boots_of_brilliance.png",
-                "of Giants": "foot/suffix/heavy_boots_of_giants.png",
-                "of Reflection": "foot/suffix/heavy_boots_of_reflection.png",
-                "of Vitriol": "foot/suffix/heavy_boots_of_vitriol.png",
-                "of Titans": "foot/suffix/heavy_boots_of_titans.png",
-                "of Skill": "foot/suffix/heavy_boots_of_skill.png",
-                "of Fury": "foot/suffix/heavy_boots_of_fury.png"
-            },
-            "Chain Boots": {
-                "of Giants": "foot/suffix/chain_boots_of_giants.png",
-                "of Perfection": "foot/suffix/chain_boots_of_perfection.png",
-                "of Reflection": "foot/suffix/chain_boots_of_reflection.png",
-                "of Enlightenment": "foot/suffix/chain_boots_of_enlightenment.png",
-                "of Detection": "foot/suffix/chain_boots_of_detection.png",
-                "of Brilliance": "foot/suffix/chain_boots_of_brilliance.png",
-                "of Vitriol": "foot/suffix/chain_boots_of_vitriol.png",
-                "of Fury": "foot/suffix/chain_boots_of_fury.png",
-                "of Protection": "foot/suffix/chain_boots_of_protection.png",
-                "of Anger": "foot/suffix/chain_boots_of_anger.png",
-                "of Power": "foot/suffix/chain_boots_of_power.png",
-                "of the Fox": "foot/suffix/chain_boots_of_the_fox.png",
-                "of Titans": "foot/suffix/chain_boots_of_titans.png",
-                "of Rage": "foot/suffix/chain_boots_of_rage.png",
-                "of the Twins": "foot/suffix/chain_boots_of_the_twins.png",
-                "of Skill": "foot/suffix/chain_boots_of_skill.png"
-            },
-            "Silk Slippers": {
-                "of Fury": "foot/suffix/silk_slippers_of_fury.png",
-                "of Reflection": "foot/suffix/silk_slippers_of_reflection.png",
-                "of Giants": "foot/suffix/silk_slippers_of_giants.png",
-                "of Brilliance": "foot/suffix/silk_slippers_of_brilliance.png",
-                "of Power": "foot/suffix/silk_slippers_of_power.png",
-                "of Enlightenment": "foot/suffix/silk_slippers_of_enlightenment.png",
-                "of the Fox": "foot/suffix/silk_slippers_of_the_fox.png",
-                "of the Twins": "foot/suffix/silk_slippers_of_the_twins.png",
-                "of Detection": "foot/suffix/silk_slippers_of_detection.png",
-                "of Vitriol": "foot/suffix/silk_slippers_of_vitriol.png",
-                "of Titans": "foot/suffix/silk_slippers_of_titans.png",
-                "of Protection": "foot/suffix/silk_slippers_of_protection.png",
-                "of Perfection": "foot/suffix/silk_slippers_of_perfection.png",
-                "of Skill": "foot/suffix/silk_slippers_of_skill.png",
-                "of Anger": "foot/suffix/silk_slippers_of_anger.png",
-                "of Rage": "foot/suffix/silk_slippers_of_rage.png"
-            },
-            "Divine Slippers": {
-                "of Giants": "foot/suffix/divine_slippers_of_giants.png",
-                "of Titans": "foot/suffix/divine_slippers_of_titans.png",
-                "of the Fox": "foot/suffix/divine_slippers_of_the_fox.png",
-                "of Perfection": "foot/suffix/divine_slippers_of_perfection.png",
-                "of the Twins": "foot/suffix/divine_slippers_of_the_twins.png",
-                "of Reflection": "foot/suffix/divine_slippers_of_reflection.png",
-                "of Detection": "foot/suffix/divine_slippers_of_detection.png",
-                "of Fury": "foot/suffix/divine_slippers_of_fury.png",
-                "of Brilliance": "foot/suffix/divine_slippers_of_brilliance.png",
-                "of Power": "foot/suffix/divine_slippers_of_power.png",
-                "of Enlightenment": "foot/suffix/divine_slippers_of_enlightenment.png",
-                "of Skill": "foot/suffix/divine_slippers_of_skill.png",
-                "of Rage": "foot/suffix/divine_slippers_of_rage.png",
-                "of Vitriol": "foot/suffix/divine_slippers_of_vitriol.png",
-                "of Anger": "foot/suffix/divine_slippers_of_anger.png",
-                "of Protection": "foot/suffix/divine_slippers_of_protection.png"
-            },
-            "Leather Boots": {
-                "of Titans": "foot/suffix/leather_boots_of_titans.png",
-                "of Detection": "foot/suffix/leather_boots_of_detection.png",
-                "of the Fox": "foot/suffix/leather_boots_of_the_fox.png",
-                "of Brilliance": "foot/suffix/leather_boots_of_brilliance.png",
-                "of Perfection": "foot/suffix/leather_boots_of_perfection.png",
-                "of Protection": "foot/suffix/leather_boots_of_protection.png",
-                "of Anger": "foot/suffix/leather_boots_of_anger.png",
-                "of Skill": "foot/suffix/leather_boots_of_skill.png",
-                "of Enlightenment": "foot/suffix/leather_boots_of_enlightenment.png",
-                "of Fury": "foot/suffix/leather_boots_of_fury.png",
-                "of Giants": "foot/suffix/leather_boots_of_giants.png",
-                "of Reflection": "foot/suffix/leather_boots_of_reflection.png",
-                "of Vitriol": "foot/suffix/leather_boots_of_vitriol.png",
-                "of Power": "foot/suffix/leather_boots_of_power.png",
-                "of the Twins": "foot/suffix/leather_boots_of_the_twins.png",
-                "of Rage": "foot/suffix/leather_boots_of_rage.png"
-            },
-            "Linen Shoes": {
-                "of Brilliance": "foot/suffix/linen_shoes_of_brilliance.png",
-                "of Enlightenment": "foot/suffix/linen_shoes_of_enlightenment.png",
-                "of Titans": "foot/suffix/linen_shoes_of_titans.png",
-                "of Perfection": "foot/suffix/linen_shoes_of_perfection.png",
-                "of Power": "foot/suffix/linen_shoes_of_power.png",
-                "of Detection": "foot/suffix/linen_shoes_of_detection.png",
-                "of Rage": "foot/suffix/linen_shoes_of_rage.png",
-                "of Vitriol": "foot/suffix/linen_shoes_of_vitriol.png",
-                "of Giants": "foot/suffix/linen_shoes_of_giants.png",
-                "of Protection": "foot/suffix/linen_shoes_of_protection.png",
-                "of the Twins": "foot/suffix/linen_shoes_of_the_twins.png",
-                "of Skill": "foot/suffix/linen_shoes_of_skill.png",
-                "of Fury": "foot/suffix/linen_shoes_of_fury.png",
-                "of Reflection": "foot/suffix/linen_shoes_of_reflection.png",
-                "of the Fox": "foot/suffix/linen_shoes_of_the_fox.png",
-                "of Anger": "foot/suffix/linen_shoes_of_anger.png"
-            },
-            "Shoes": {
-                "of Giants": "foot/suffix/shoes_of_giants.png",
-                "of Perfection": "foot/suffix/shoes_of_perfection.png",
-                "of Skill": "foot/suffix/shoes_of_skill.png",
-                "of Anger": "foot/suffix/shoes_of_anger.png",
-                "of Enlightenment": "foot/suffix/shoes_of_enlightenment.png",
-                "of Vitriol": "foot/suffix/shoes_of_vitriol.png",
-                "of Titans": "foot/suffix/shoes_of_titans.png",
-                "of Fury": "foot/suffix/shoes_of_fury.png",
-                "of the Fox": "foot/suffix/shoes_of_the_fox.png",
-                "of Brilliance": "foot/suffix/shoes_of_brilliance.png",
-                "of Rage": "foot/suffix/shoes_of_rage.png",
-                "of Power": "foot/suffix/shoes_of_power.png",
-                "of Reflection": "foot/suffix/shoes_of_reflection.png",
-                "of Detection": "foot/suffix/shoes_of_detection.png",
-                "of Protection": "foot/suffix/shoes_of_protection.png",
-                "of the Twins": "foot/suffix/shoes_of_the_twins.png"
-            }
-        },
-        "plus_one": "foot/plus_one/plus_one.png",
-        "name": {
-            "Ornate Greaves": "foot/name/ornate_greaves.png",
-            "Demonhide Boots": "foot/name/demonhide_boots.png",
-            "Holy Greaves": "foot/name/holy_greaves.png",
-            "Dragonskin Boots": "foot/name/dragonskin_boots.png",
-            "Greaves": "foot/name/greaves.png",
-            "Hard Leather Boots": "foot/name/hard_leather_boots.png",
-            "Wool Shoes": "foot/name/wool_shoes.png",
-            "Studded Leather Boots": "foot/name/studded_leather_boots.png",
-            "Heavy Boots": "foot/name/heavy_boots.png",
-            "Chain Boots": "foot/name/chain_boots.png",
-            "Silk Slippers": "foot/name/silk_slippers.png",
-            "Divine Slippers": "foot/name/divine_slippers.png",
-            "Leather Boots": "foot/name/leather_boots.png",
-            "Linen Shoes": "foot/name/linen_shoes.png",
-            "Shoes": "foot/name/shoes.png"
-        }
-    },
-    "hand": {
-        "prefix": "hand/prefix/prefix.png",
-        "suffix": {
-            "Studded Leather Gloves": {
-                "of the Twins": "hand/suffix/studded_leather_gloves_of_the_twins.png",
-                "of Brilliance": "hand/suffix/studded_leather_gloves_of_brilliance.png",
-                "of Fury": "hand/suffix/studded_leather_gloves_of_fury.png",
-                "of Reflection": "hand/suffix/studded_leather_gloves_of_reflection.png",
-                "of Detection": "hand/suffix/studded_leather_gloves_of_detection.png",
-                "of Rage": "hand/suffix/studded_leather_gloves_of_rage.png",
-                "of the Fox": "hand/suffix/studded_leather_gloves_of_the_fox.png",
-                "of Perfection": "hand/suffix/studded_leather_gloves_of_perfection.png",
-                "of Titans": "hand/suffix/studded_leather_gloves_of_titans.png",
-                "of Anger": "hand/suffix/studded_leather_gloves_of_anger.png",
-                "of Skill": "hand/suffix/studded_leather_gloves_of_skill.png",
-                "of Protection": "hand/suffix/studded_leather_gloves_of_protection.png",
-                "of Giants": "hand/suffix/studded_leather_gloves_of_giants.png",
-                "of Enlightenment": "hand/suffix/studded_leather_gloves_of_enlightenment.png",
-                "of Vitriol": "hand/suffix/studded_leather_gloves_of_vitriol.png",
-                "of Power": "hand/suffix/studded_leather_gloves_of_power.png"
-            },
-            "Gloves": {
-                "of Anger": "hand/suffix/gloves_of_anger.png",
-                "of Skill": "hand/suffix/gloves_of_skill.png",
-                "of Enlightenment": "hand/suffix/gloves_of_enlightenment.png",
-                "of Rage": "hand/suffix/gloves_of_rage.png",
-                "of Titans": "hand/suffix/gloves_of_titans.png",
-                "of Perfection": "hand/suffix/gloves_of_perfection.png",
-                "of Giants": "hand/suffix/gloves_of_giants.png",
-                "of Brilliance": "hand/suffix/gloves_of_brilliance.png",
-                "of Detection": "hand/suffix/gloves_of_detection.png",
-                "of Power": "hand/suffix/gloves_of_power.png",
-                "of Protection": "hand/suffix/gloves_of_protection.png",
-                "of Reflection": "hand/suffix/gloves_of_reflection.png",
-                "of the Twins": "hand/suffix/gloves_of_the_twins.png",
-                "of Fury": "hand/suffix/gloves_of_fury.png",
-                "of the Fox": "hand/suffix/gloves_of_the_fox.png",
-                "of Vitriol": "hand/suffix/gloves_of_vitriol.png"
-            },
-            "Dragonskin Gloves": {
-                "of Power": "hand/suffix/dragonskin_gloves_of_power.png",
-                "of Fury": "hand/suffix/dragonskin_gloves_of_fury.png",
-                "of the Twins": "hand/suffix/dragonskin_gloves_of_the_twins.png",
-                "of Protection": "hand/suffix/dragonskin_gloves_of_protection.png",
-                "of Vitriol": "hand/suffix/dragonskin_gloves_of_vitriol.png",
-                "of Perfection": "hand/suffix/dragonskin_gloves_of_perfection.png",
-                "of Anger": "hand/suffix/dragonskin_gloves_of_anger.png",
-                "of Giants": "hand/suffix/dragonskin_gloves_of_giants.png",
-                "of Brilliance": "hand/suffix/dragonskin_gloves_of_brilliance.png",
-                "of Detection": "hand/suffix/dragonskin_gloves_of_detection.png",
-                "of Rage": "hand/suffix/dragonskin_gloves_of_rage.png",
-                "of Reflection": "hand/suffix/dragonskin_gloves_of_reflection.png",
-                "of Titans": "hand/suffix/dragonskin_gloves_of_titans.png",
-                "of Enlightenment": "hand/suffix/dragonskin_gloves_of_enlightenment.png",
-                "of Skill": "hand/suffix/dragonskin_gloves_of_skill.png",
-                "of the Fox": "hand/suffix/dragonskin_gloves_of_the_fox.png"
-            },
-            "Leather Gloves": {
-                "of Perfection": "hand/suffix/leather_gloves_of_perfection.png",
-                "of Reflection": "hand/suffix/leather_gloves_of_reflection.png",
-                "of Enlightenment": "hand/suffix/leather_gloves_of_enlightenment.png",
-                "of the Fox": "hand/suffix/leather_gloves_of_the_fox.png",
-                "of Rage": "hand/suffix/leather_gloves_of_rage.png",
-                "of Brilliance": "hand/suffix/leather_gloves_of_brilliance.png",
-                "of Detection": "hand/suffix/leather_gloves_of_detection.png",
-                "of Anger": "hand/suffix/leather_gloves_of_anger.png",
-                "of Fury": "hand/suffix/leather_gloves_of_fury.png",
-                "of the Twins": "hand/suffix/leather_gloves_of_the_twins.png",
-                "of Power": "hand/suffix/leather_gloves_of_power.png",
-                "of Protection": "hand/suffix/leather_gloves_of_protection.png",
-                "of Skill": "hand/suffix/leather_gloves_of_skill.png",
-                "of Giants": "hand/suffix/leather_gloves_of_giants.png",
-                "of Titans": "hand/suffix/leather_gloves_of_titans.png",
-                "of Vitriol": "hand/suffix/leather_gloves_of_vitriol.png"
-            },
-            "Hard Leather Gloves": {
-                "of Skill": "hand/suffix/hard_leather_gloves_of_skill.png",
-                "of the Twins": "hand/suffix/hard_leather_gloves_of_the_twins.png",
-                "of the Fox": "hand/suffix/hard_leather_gloves_of_the_fox.png",
-                "of Titans": "hand/suffix/hard_leather_gloves_of_titans.png",
-                "of Reflection": "hand/suffix/hard_leather_gloves_of_reflection.png",
-                "of Anger": "hand/suffix/hard_leather_gloves_of_anger.png",
-                "of Power": "hand/suffix/hard_leather_gloves_of_power.png",
-                "of Vitriol": "hand/suffix/hard_leather_gloves_of_vitriol.png",
-                "of Protection": "hand/suffix/hard_leather_gloves_of_protection.png",
-                "of Rage": "hand/suffix/hard_leather_gloves_of_rage.png",
-                "of Giants": "hand/suffix/hard_leather_gloves_of_giants.png",
-                "of Enlightenment": "hand/suffix/hard_leather_gloves_of_enlightenment.png",
-                "of Fury": "hand/suffix/hard_leather_gloves_of_fury.png",
-                "of Perfection": "hand/suffix/hard_leather_gloves_of_perfection.png",
-                "of Detection": "hand/suffix/hard_leather_gloves_of_detection.png",
-                "of Brilliance": "hand/suffix/hard_leather_gloves_of_brilliance.png"
-            },
-            "Heavy Gloves": {
-                "of Titans": "hand/suffix/heavy_gloves_of_titans.png",
-                "of Skill": "hand/suffix/heavy_gloves_of_skill.png",
-                "of Fury": "hand/suffix/heavy_gloves_of_fury.png",
-                "of the Twins": "hand/suffix/heavy_gloves_of_the_twins.png",
-                "of Detection": "hand/suffix/heavy_gloves_of_detection.png",
-                "of Brilliance": "hand/suffix/heavy_gloves_of_brilliance.png",
-                "of Enlightenment": "hand/suffix/heavy_gloves_of_enlightenment.png",
-                "of Vitriol": "hand/suffix/heavy_gloves_of_vitriol.png",
-                "of the Fox": "hand/suffix/heavy_gloves_of_the_fox.png",
-                "of Protection": "hand/suffix/heavy_gloves_of_protection.png",
-                "of Power": "hand/suffix/heavy_gloves_of_power.png",
-                "of Perfection": "hand/suffix/heavy_gloves_of_perfection.png",
-                "of Rage": "hand/suffix/heavy_gloves_of_rage.png",
-                "of Giants": "hand/suffix/heavy_gloves_of_giants.png",
-                "of Anger": "hand/suffix/heavy_gloves_of_anger.png",
-                "of Reflection": "hand/suffix/heavy_gloves_of_reflection.png"
-            },
-            "Gauntlets": {
-                "of the Fox": "hand/suffix/gauntlets_of_the_fox.png",
-                "of Detection": "hand/suffix/gauntlets_of_detection.png",
-                "of Rage": "hand/suffix/gauntlets_of_rage.png",
-                "of the Twins": "hand/suffix/gauntlets_of_the_twins.png",
-                "of Reflection": "hand/suffix/gauntlets_of_reflection.png",
-                "of Vitriol": "hand/suffix/gauntlets_of_vitriol.png",
-                "of Titans": "hand/suffix/gauntlets_of_titans.png",
-                "of Fury": "hand/suffix/gauntlets_of_fury.png",
-                "of Enlightenment": "hand/suffix/gauntlets_of_enlightenment.png",
-                "of Giants": "hand/suffix/gauntlets_of_giants.png",
-                "of Skill": "hand/suffix/gauntlets_of_skill.png",
-                "of Power": "hand/suffix/gauntlets_of_power.png",
-                "of Anger": "hand/suffix/gauntlets_of_anger.png",
-                "of Perfection": "hand/suffix/gauntlets_of_perfection.png",
-                "of Protection": "hand/suffix/gauntlets_of_protection.png",
-                "of Brilliance": "hand/suffix/gauntlets_of_brilliance.png"
-            },
-            "Wool Gloves": {
-                "of Skill": "hand/suffix/wool_gloves_of_skill.png",
-                "of Anger": "hand/suffix/wool_gloves_of_anger.png",
-                "of Rage": "hand/suffix/wool_gloves_of_rage.png",
-                "of Brilliance": "hand/suffix/wool_gloves_of_brilliance.png",
-                "of Reflection": "hand/suffix/wool_gloves_of_reflection.png",
-                "of Titans": "hand/suffix/wool_gloves_of_titans.png",
-                "of Detection": "hand/suffix/wool_gloves_of_detection.png",
-                "of Perfection": "hand/suffix/wool_gloves_of_perfection.png",
-                "of Protection": "hand/suffix/wool_gloves_of_protection.png",
-                "of Vitriol": "hand/suffix/wool_gloves_of_vitriol.png",
-                "of Enlightenment": "hand/suffix/wool_gloves_of_enlightenment.png",
-                "of the Twins": "hand/suffix/wool_gloves_of_the_twins.png",
-                "of the Fox": "hand/suffix/wool_gloves_of_the_fox.png",
-                "of Power": "hand/suffix/wool_gloves_of_power.png",
-                "of Giants": "hand/suffix/wool_gloves_of_giants.png",
-                "of Fury": "hand/suffix/wool_gloves_of_fury.png"
-            },
-            "Silk Gloves": {
-                "of Detection": "hand/suffix/silk_gloves_of_detection.png",
-                "of the Twins": "hand/suffix/silk_gloves_of_the_twins.png",
-                "of Brilliance": "hand/suffix/silk_gloves_of_brilliance.png",
-                "of Titans": "hand/suffix/silk_gloves_of_titans.png",
-                "of Perfection": "hand/suffix/silk_gloves_of_perfection.png",
-                "of Skill": "hand/suffix/silk_gloves_of_skill.png",
-                "of Anger": "hand/suffix/silk_gloves_of_anger.png",
-                "of the Fox": "hand/suffix/silk_gloves_of_the_fox.png",
-                "of Enlightenment": "hand/suffix/silk_gloves_of_enlightenment.png",
-                "of Reflection": "hand/suffix/silk_gloves_of_reflection.png",
-                "of Fury": "hand/suffix/silk_gloves_of_fury.png",
-                "of Vitriol": "hand/suffix/silk_gloves_of_vitriol.png",
-                "of Giants": "hand/suffix/silk_gloves_of_giants.png",
-                "of Power": "hand/suffix/silk_gloves_of_power.png",
-                "of Protection": "hand/suffix/silk_gloves_of_protection.png",
-                "of Rage": "hand/suffix/silk_gloves_of_rage.png"
-            },
-            "Ornate Gauntlets": {
-                "of Vitriol": "hand/suffix/ornate_gauntlets_of_vitriol.png",
-                "of Titans": "hand/suffix/ornate_gauntlets_of_titans.png",
-                "of Brilliance": "hand/suffix/ornate_gauntlets_of_brilliance.png",
-                "of Protection": "hand/suffix/ornate_gauntlets_of_protection.png",
-                "of Anger": "hand/suffix/ornate_gauntlets_of_anger.png",
-                "of Enlightenment": "hand/suffix/ornate_gauntlets_of_enlightenment.png",
-                "of Rage": "hand/suffix/ornate_gauntlets_of_rage.png",
-                "of Giants": "hand/suffix/ornate_gauntlets_of_giants.png",
-                "of the Twins": "hand/suffix/ornate_gauntlets_of_the_twins.png",
-                "of Reflection": "hand/suffix/ornate_gauntlets_of_reflection.png",
-                "of Skill": "hand/suffix/ornate_gauntlets_of_skill.png",
-                "of Power": "hand/suffix/ornate_gauntlets_of_power.png",
-                "of Fury": "hand/suffix/ornate_gauntlets_of_fury.png",
-                "of Perfection": "hand/suffix/ornate_gauntlets_of_perfection.png",
-                "of Detection": "hand/suffix/ornate_gauntlets_of_detection.png",
-                "of the Fox": "hand/suffix/ornate_gauntlets_of_the_fox.png"
-            },
-            "Demon's Hands": {
-                "of Vitriol": "hand/suffix/demons_hands_of_vitriol.png",
-                "of Skill": "hand/suffix/demons_hands_of_skill.png",
-                "of Anger": "hand/suffix/demons_hands_of_anger.png",
-                "of Brilliance": "hand/suffix/demons_hands_of_brilliance.png",
-                "of Reflection": "hand/suffix/demons_hands_of_reflection.png",
-                "of Detection": "hand/suffix/demons_hands_of_detection.png",
-                "of Power": "hand/suffix/demons_hands_of_power.png",
-                "of Giants": "hand/suffix/demons_hands_of_giants.png",
-                "of the Fox": "hand/suffix/demons_hands_of_the_fox.png",
-                "of Titans": "hand/suffix/demons_hands_of_titans.png",
-                "of Enlightenment": "hand/suffix/demons_hands_of_enlightenment.png",
-                "of Perfection": "hand/suffix/demons_hands_of_perfection.png",
-                "of Fury": "hand/suffix/demons_hands_of_fury.png",
-                "of Rage": "hand/suffix/demons_hands_of_rage.png",
-                "of Protection": "hand/suffix/demons_hands_of_protection.png",
-                "of the Twins": "hand/suffix/demons_hands_of_the_twins.png"
-            },
-            "Divine Gloves": {
-                "of Anger": "hand/suffix/divine_gloves_of_anger.png",
-                "of Brilliance": "hand/suffix/divine_gloves_of_brilliance.png",
-                "of the Fox": "hand/suffix/divine_gloves_of_the_fox.png",
-                "of Vitriol": "hand/suffix/divine_gloves_of_vitriol.png",
-                "of Skill": "hand/suffix/divine_gloves_of_skill.png",
-                "of the Twins": "hand/suffix/divine_gloves_of_the_twins.png",
-                "of Detection": "hand/suffix/divine_gloves_of_detection.png",
-                "of Enlightenment": "hand/suffix/divine_gloves_of_enlightenment.png",
-                "of Power": "hand/suffix/divine_gloves_of_power.png",
-                "of Titans": "hand/suffix/divine_gloves_of_titans.png",
-                "of Giants": "hand/suffix/divine_gloves_of_giants.png",
-                "of Perfection": "hand/suffix/divine_gloves_of_perfection.png",
-                "of Fury": "hand/suffix/divine_gloves_of_fury.png",
-                "of Protection": "hand/suffix/divine_gloves_of_protection.png",
-                "of Reflection": "hand/suffix/divine_gloves_of_reflection.png",
-                "of Rage": "hand/suffix/divine_gloves_of_rage.png"
-            },
-            "Linen Gloves": {
-                "of Titans": "hand/suffix/linen_gloves_of_titans.png",
-                "of Vitriol": "hand/suffix/linen_gloves_of_vitriol.png",
-                "of Reflection": "hand/suffix/linen_gloves_of_reflection.png",
-                "of Skill": "hand/suffix/linen_gloves_of_skill.png",
-                "of the Fox": "hand/suffix/linen_gloves_of_the_fox.png",
-                "of the Twins": "hand/suffix/linen_gloves_of_the_twins.png",
-                "of Perfection": "hand/suffix/linen_gloves_of_perfection.png",
-                "of Detection": "hand/suffix/linen_gloves_of_detection.png",
-                "of Power": "hand/suffix/linen_gloves_of_power.png",
-                "of Brilliance": "hand/suffix/linen_gloves_of_brilliance.png",
-                "of Protection": "hand/suffix/linen_gloves_of_protection.png",
-                "of Anger": "hand/suffix/linen_gloves_of_anger.png",
-                "of Fury": "hand/suffix/linen_gloves_of_fury.png",
-                "of Giants": "hand/suffix/linen_gloves_of_giants.png",
-                "of Rage": "hand/suffix/linen_gloves_of_rage.png",
-                "of Enlightenment": "hand/suffix/linen_gloves_of_enlightenment.png"
-            },
-            "Holy Gauntlets": {
-                "of Enlightenment": "hand/suffix/holy_gauntlets_of_enlightenment.png",
-                "of Giants": "hand/suffix/holy_gauntlets_of_giants.png",
-                "of Vitriol": "hand/suffix/holy_gauntlets_of_vitriol.png",
-                "of Power": "hand/suffix/holy_gauntlets_of_power.png",
-                "of Perfection": "hand/suffix/holy_gauntlets_of_perfection.png",
-                "of Brilliance": "hand/suffix/holy_gauntlets_of_brilliance.png",
-                "of Rage": "hand/suffix/holy_gauntlets_of_rage.png",
-                "of Fury": "hand/suffix/holy_gauntlets_of_fury.png",
-                "of Titans": "hand/suffix/holy_gauntlets_of_titans.png",
-                "of the Fox": "hand/suffix/holy_gauntlets_of_the_fox.png",
-                "of Anger": "hand/suffix/holy_gauntlets_of_anger.png",
-                "of Protection": "hand/suffix/holy_gauntlets_of_protection.png",
-                "of the Twins": "hand/suffix/holy_gauntlets_of_the_twins.png",
-                "of Detection": "hand/suffix/holy_gauntlets_of_detection.png",
-                "of Reflection": "hand/suffix/holy_gauntlets_of_reflection.png",
-                "of Skill": "hand/suffix/holy_gauntlets_of_skill.png"
-            },
-            "Chain Gloves": {
-                "of Fury": "hand/suffix/chain_gloves_of_fury.png",
-                "of Power": "hand/suffix/chain_gloves_of_power.png",
-                "of Brilliance": "hand/suffix/chain_gloves_of_brilliance.png",
-                "of Anger": "hand/suffix/chain_gloves_of_anger.png",
-                "of Detection": "hand/suffix/chain_gloves_of_detection.png",
-                "of Enlightenment": "hand/suffix/chain_gloves_of_enlightenment.png",
-                "of Titans": "hand/suffix/chain_gloves_of_titans.png",
-                "of Rage": "hand/suffix/chain_gloves_of_rage.png",
-                "of Giants": "hand/suffix/chain_gloves_of_giants.png",
-                "of Vitriol": "hand/suffix/chain_gloves_of_vitriol.png",
-                "of the Twins": "hand/suffix/chain_gloves_of_the_twins.png",
-                "of Perfection": "hand/suffix/chain_gloves_of_perfection.png",
-                "of the Fox": "hand/suffix/chain_gloves_of_the_fox.png",
-                "of Skill": "hand/suffix/chain_gloves_of_skill.png",
-                "of Protection": "hand/suffix/chain_gloves_of_protection.png",
-                "of Reflection": "hand/suffix/chain_gloves_of_reflection.png"
-            }
-        },
-        "plus_one": "hand/plus_one/plus_one.png",
-        "name": {
-            "Studded Leather Gloves": "hand/name/studded_leather_gloves.png",
-            "Gloves": "hand/name/gloves.png",
-            "Dragonskin Gloves": "hand/name/dragonskin_gloves.png",
-            "Leather Gloves": "hand/name/leather_gloves.png",
-            "Hard Leather Gloves": "hand/name/hard_leather_gloves.png",
-            "Heavy Gloves": "hand/name/heavy_gloves.png",
-            "Gauntlets": "hand/name/gauntlets.png",
-            "Wool Gloves": "hand/name/wool_gloves.png",
-            "Silk Gloves": "hand/name/silk_gloves.png",
-            "Ornate Gauntlets": "hand/name/ornate_gauntlets.png",
-            "Demon's Hands": "hand/name/demons_hands.png",
-            "Divine Gloves": "hand/name/divine_gloves.png",
-            "Linen Gloves": "hand/name/linen_gloves.png",
-            "Holy Gauntlets": "hand/name/holy_gauntlets.png",
-            "Chain Gloves": "hand/name/chain_gloves.png"
-        }
-    },
-    "neck": {
-        "prefix": "neck/prefix/prefix.png",
-        "suffix": {
-            "Necklace": {
-                "of Enlightenment": "neck/suffix/necklace_of_enlightenment.png",
-                "of Detection": "neck/suffix/necklace_of_detection.png",
-                "of the Twins": "neck/suffix/necklace_of_the_twins.png",
-                "of Fury": "neck/suffix/necklace_of_fury.png",
-                "of Protection": "neck/suffix/necklace_of_protection.png",
-                "of the Fox": "neck/suffix/necklace_of_the_fox.png",
-                "of Brilliance": "neck/suffix/necklace_of_brilliance.png",
-                "of Rage": "neck/suffix/necklace_of_rage.png",
-                "of Vitriol": "neck/suffix/necklace_of_vitriol.png",
-                "of Titans": "neck/suffix/necklace_of_titans.png",
-                "of Power": "neck/suffix/necklace_of_power.png",
-                "of Skill": "neck/suffix/necklace_of_skill.png",
-                "of Perfection": "neck/suffix/necklace_of_perfection.png",
-                "of Giants": "neck/suffix/necklace_of_giants.png",
-                "of Reflection": "neck/suffix/necklace_of_reflection.png",
-                "of Anger": "neck/suffix/necklace_of_anger.png"
-            },
-            "Amulet": {
-                "of Reflection": "neck/suffix/amulet_of_reflection.png",
-                "of Enlightenment": "neck/suffix/amulet_of_enlightenment.png",
-                "of Power": "neck/suffix/amulet_of_power.png",
-                "of Rage": "neck/suffix/amulet_of_rage.png",
-                "of Vitriol": "neck/suffix/amulet_of_vitriol.png",
-                "of the Fox": "neck/suffix/amulet_of_the_fox.png",
-                "of Perfection": "neck/suffix/amulet_of_perfection.png",
-                "of Detection": "neck/suffix/amulet_of_detection.png",
-                "of Fury": "neck/suffix/amulet_of_fury.png",
-                "of Anger": "neck/suffix/amulet_of_anger.png",
-                "of Skill": "neck/suffix/amulet_of_skill.png",
-                "of the Twins": "neck/suffix/amulet_of_the_twins.png",
-                "of Titans": "neck/suffix/amulet_of_titans.png",
-                "of Protection": "neck/suffix/amulet_of_protection.png",
-                "of Brilliance": "neck/suffix/amulet_of_brilliance.png",
-                "of Giants": "neck/suffix/amulet_of_giants.png"
-            },
-            "Pendant": {
-                "of Brilliance": "neck/suffix/pendant_of_brilliance.png",
-                "of Fury": "neck/suffix/pendant_of_fury.png",
-                "of Detection": "neck/suffix/pendant_of_detection.png",
-                "of Power": "neck/suffix/pendant_of_power.png",
-                "of Rage": "neck/suffix/pendant_of_rage.png",
-                "of Vitriol": "neck/suffix/pendant_of_vitriol.png",
-                "of Skill": "neck/suffix/pendant_of_skill.png",
-                "of the Fox": "neck/suffix/pendant_of_the_fox.png",
-                "of Titans": "neck/suffix/pendant_of_titans.png",
-                "of Reflection": "neck/suffix/pendant_of_reflection.png",
-                "of the Twins": "neck/suffix/pendant_of_the_twins.png",
-                "of Giants": "neck/suffix/pendant_of_giants.png",
-                "of Protection": "neck/suffix/pendant_of_protection.png",
-                "of Enlightenment": "neck/suffix/pendant_of_enlightenment.png",
-                "of Perfection": "neck/suffix/pendant_of_perfection.png",
-                "of Anger": "neck/suffix/pendant_of_anger.png"
-            }
-        },
-        "plus_one": "neck/plus_one/plus_one.png",
-        "name": {
-            "Necklace": "neck/name/necklace.png",
-            "Amulet": "neck/name/amulet.png",
-            "Pendant": "neck/name/pendant.png"
-        }
-    },
-    "ring": {
-        "prefix": "ring/prefix/prefix.png",
-        "suffix": {
-            "Gold Ring": {
-                "of Titans": "ring/suffix/gold_ring_of_titans.png",
-                "of Anger": "ring/suffix/gold_ring_of_anger.png",
-                "of Rage": "ring/suffix/gold_ring_of_rage.png",
-                "of Perfection": "ring/suffix/gold_ring_of_perfection.png",
-                "of Enlightenment": "ring/suffix/gold_ring_of_enlightenment.png",
-                "of Reflection": "ring/suffix/gold_ring_of_reflection.png",
-                "of Fury": "ring/suffix/gold_ring_of_fury.png",
-                "of the Fox": "ring/suffix/gold_ring_of_the_fox.png",
-                "of Vitriol": "ring/suffix/gold_ring_of_vitriol.png",
-                "of the Twins": "ring/suffix/gold_ring_of_the_twins.png",
-                "of Giants": "ring/suffix/gold_ring_of_giants.png",
-                "of Protection": "ring/suffix/gold_ring_of_protection.png",
-                "of Detection": "ring/suffix/gold_ring_of_detection.png",
-                "of Skill": "ring/suffix/gold_ring_of_skill.png",
-                "of Brilliance": "ring/suffix/gold_ring_of_brilliance.png",
-                "of Power": "ring/suffix/gold_ring_of_power.png"
-            },
-            "Titanium Ring": {
-                "of Brilliance": "ring/suffix/titanium_ring_of_brilliance.png",
-                "of Anger": "ring/suffix/titanium_ring_of_anger.png",
-                "of Titans": "ring/suffix/titanium_ring_of_titans.png",
-                "of Power": "ring/suffix/titanium_ring_of_power.png",
-                "of Giants": "ring/suffix/titanium_ring_of_giants.png",
-                "of Detection": "ring/suffix/titanium_ring_of_detection.png",
-                "of Skill": "ring/suffix/titanium_ring_of_skill.png",
-                "of Fury": "ring/suffix/titanium_ring_of_fury.png",
-                "of Enlightenment": "ring/suffix/titanium_ring_of_enlightenment.png",
-                "of Rage": "ring/suffix/titanium_ring_of_rage.png",
-                "of Reflection": "ring/suffix/titanium_ring_of_reflection.png",
-                "of Vitriol": "ring/suffix/titanium_ring_of_vitriol.png",
-                "of the Fox": "ring/suffix/titanium_ring_of_the_fox.png",
-                "of Protection": "ring/suffix/titanium_ring_of_protection.png",
-                "of the Twins": "ring/suffix/titanium_ring_of_the_twins.png",
-                "of Perfection": "ring/suffix/titanium_ring_of_perfection.png"
-            },
-            "Silver Ring": {
-                "of the Twins": "ring/suffix/silver_ring_of_the_twins.png",
-                "of Reflection": "ring/suffix/silver_ring_of_reflection.png",
-                "of Skill": "ring/suffix/silver_ring_of_skill.png",
-                "of Anger": "ring/suffix/silver_ring_of_anger.png",
-                "of Fury": "ring/suffix/silver_ring_of_fury.png",
-                "of Giants": "ring/suffix/silver_ring_of_giants.png",
-                "of Protection": "ring/suffix/silver_ring_of_protection.png",
-                "of Perfection": "ring/suffix/silver_ring_of_perfection.png",
-                "of Vitriol": "ring/suffix/silver_ring_of_vitriol.png",
-                "of Power": "ring/suffix/silver_ring_of_power.png",
-                "of Titans": "ring/suffix/silver_ring_of_titans.png",
-                "of Detection": "ring/suffix/silver_ring_of_detection.png",
-                "of the Fox": "ring/suffix/silver_ring_of_the_fox.png",
-                "of Enlightenment": "ring/suffix/silver_ring_of_enlightenment.png",
-                "of Rage": "ring/suffix/silver_ring_of_rage.png",
-                "of Brilliance": "ring/suffix/silver_ring_of_brilliance.png"
-            },
-            "Platinum Ring": {
-                "of Detection": "ring/suffix/platinum_ring_of_detection.png",
-                "of Power": "ring/suffix/platinum_ring_of_power.png",
-                "of Enlightenment": "ring/suffix/platinum_ring_of_enlightenment.png",
-                "of Rage": "ring/suffix/platinum_ring_of_rage.png",
-                "of Reflection": "ring/suffix/platinum_ring_of_reflection.png",
-                "of Vitriol": "ring/suffix/platinum_ring_of_vitriol.png",
-                "of Giants": "ring/suffix/platinum_ring_of_giants.png",
-                "of the Fox": "ring/suffix/platinum_ring_of_the_fox.png",
-                "of Skill": "ring/suffix/platinum_ring_of_skill.png",
-                "of Fury": "ring/suffix/platinum_ring_of_fury.png",
-                "of the Twins": "ring/suffix/platinum_ring_of_the_twins.png",
-                "of Titans": "ring/suffix/platinum_ring_of_titans.png",
-                "of Brilliance": "ring/suffix/platinum_ring_of_brilliance.png",
-                "of Anger": "ring/suffix/platinum_ring_of_anger.png",
-                "of Perfection": "ring/suffix/platinum_ring_of_perfection.png",
-                "of Protection": "ring/suffix/platinum_ring_of_protection.png"
-            },
-            "Bronze Ring": {
-                "of the Twins": "ring/suffix/bronze_ring_of_the_twins.png",
-                "of Protection": "ring/suffix/bronze_ring_of_protection.png",
-                "of Skill": "ring/suffix/bronze_ring_of_skill.png",
-                "of Anger": "ring/suffix/bronze_ring_of_anger.png",
-                "of Power": "ring/suffix/bronze_ring_of_power.png",
-                "of Giants": "ring/suffix/bronze_ring_of_giants.png",
-                "of Fury": "ring/suffix/bronze_ring_of_fury.png",
-                "of the Fox": "ring/suffix/bronze_ring_of_the_fox.png",
-                "of Perfection": "ring/suffix/bronze_ring_of_perfection.png",
-                "of Titans": "ring/suffix/bronze_ring_of_titans.png",
-                "of Reflection": "ring/suffix/bronze_ring_of_reflection.png",
-                "of Detection": "ring/suffix/bronze_ring_of_detection.png",
-                "of Enlightenment": "ring/suffix/bronze_ring_of_enlightenment.png",
-                "of Brilliance": "ring/suffix/bronze_ring_of_brilliance.png",
-                "of Vitriol": "ring/suffix/bronze_ring_of_vitriol.png",
-                "of Rage": "ring/suffix/bronze_ring_of_rage.png"
-            }
-        },
-        "plus_one": "ring/plus_one/plus_one.png",
-        "name": {
-            "Gold Ring": "ring/name/gold_ring.png",
-            "Titanium Ring": "ring/name/titanium_ring.png",
-            "Silver Ring": "ring/name/silver_ring.png",
-            "Platinum Ring": "ring/name/platinum_ring.png",
-            "Bronze Ring": "ring/name/bronze_ring.png"
-        }
-    }
-}
