--- dataverse/apis/artifacts/layerOrder.ts
@@ -1,36 +1,0 @@
-export const LAYER_ORDER = [
-    "bg",
-    "fg",
-    "weaponName",
-    "weaponPrefix",
-    "weaponSuffix",
-    "weaponPlusOne",
-    "chestName",
-    "chestPrefix",
-    "chestSuffix",
-    "chestPlusOne",
-    "headName",
-    "headPrefix",
-    "headSuffix",
-    "headPlusOne",
-    "waistName",
-    "waistPrefix",
-    "waistSuffix",
-    "waistPlusOne",
-    "footName",
-    "footPrefix",
-    "footSuffix",
-    "footPlusOne",
-    "handName",
-    "handPrefix",
-    "handSuffix",
-    "handPlusOne",
-    "neckName",
-    "neckPrefix",
-    "neckSuffix",
-    "neckPlusOne",
-    "ringName",
-    "ringPrefix",
-    "ringSuffix",
-    "ringPlusOne"
-]
